% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTransitions.R
\name{calculateTransitions}
\alias{calculateTransitions}
\title{Find trees at NEON sites whose status has changed between sampling events}
\usage{
calculateTransitions(biomassTable, plotYearTable)
}
\arguments{
\item{biomassTable}{Table (generated in estimateWoodProd()) containing biomass records for individual trees. \link{data.frame}}

\item{plotYearTable}{The 'vst_perplotperyear' table for the site x month combination(s) of interest; this table is passed by the user into the estimateWoodProd() function. \link{data.frame}}
}
\value{
A table with one record for each individual whose status has changed, noting its status and biomass at each time point.
}
\description{
Identifies individual trees whose status has changed between surveys (e.g. live to dead), and infers the appropriate status for missed events when possible; for example, when a tree is "live" then is not sampled for several eventIDs and is "live" again, it is inferred to be alive for the eventIDs that were missed. Biomass is also inferred via interpolation for missed eventIDs but biomass is not inferred for missed eventIDs outside the recorded data (i.e., no extrapolation). The growthInterval is then determined using dates for both observed and inferred biomass.
}
\details{
This is a helper function for the estimateWoodProd() function, used to estimate mortality, recruitment, and missingness.
}
\author{
Courtney L Meier \email{cmeier@BattelleEcology.org} \cr
Claire K Lunch \email{clunch@BattelleEcology.org} \cr
}
\keyword{internal}
