% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divStack.R
\name{divStack}
\alias{divStack}
\title{Stack NEON plant diversity data}
\usage{
divStack(data_1m2 = NA, data_10_100m2 = NA)
}
\arguments{
\item{data_1m2}{div_1m2Data table in NEON data produdct (Plant presence and percent cover, DP1.10058.001)}

\item{data_10_100m2}{div_10m2Data100m2Data table in NEON data produdct (Plant presence and percent cover, DP1.10058.001)}
}
\value{
This function returns a data frame
}
\description{
Use this function to aggregate NEON plant presence data to reflect plant species present at each plot scale.
}
\examples{
\dontrun{

# load neonUtilities
library (neonUtilities)

# get data
allDiv <- loadByProduct(
  dpID = "DP1.10058.001",
  site = "SRER",
  startdate = "2010-01",
  package = "basic",
  check.size = FALSE)

##extract 1m2 data from list of lists and some processing
data_1m2 <- allDiv[["div_1m2Data"]]
###get 10_100
data_10_100m2 <- allDiv[["div_10m2Data100m2Data"]]

data_stacked <- divStack(
  data_1m2 = data_1m2,
  data_10_100m2 = data_10_100m2)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Dave T Barnett \email{dbarnettl@battelleecology.org} \cr
}
