% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAllometricOtherMass.R
\name{estimateAllometricOtherMass}
\alias{estimateAllometricOtherMass}
\title{Estimate mass of non-woody individuals with allometric equations}
\usage{
estimateAllometricOtherMass(nonWoodyTable, growthFormSubset = NA)
}
\arguments{
\item{nonWoodyTable}{Table (generated in estimateWoodMass()) containing records for qualifying stems of non-woody individuals.}

\item{growthFormSubset}{Select Vegetation Structure growth forms for biomass estimation. The options are "tree", which enables biomass estimation only for palm trees and large tree ferns with a DBH ≥ 10 cm, and the default of "all", which includes "tree" individuals, and also small palms, small tree ferns, ferns, ocotillo, and xerophyllum individuals. \link{character}}
}
\value{
A table with biomass estimates:
\itemize{
\item For palm trees, tree-type small palms, large tree ferns, and small tree ferns: One record for each qualifying bole/stem at each time point it was measured.
\item For shrub-type small palms, ocotillo, and xerophyllum: One record for each qualifying individual at each time point it was measured.
}
}
\description{
Estimate the biomass of non-woody individuals via growthForm-specific, and species-specific allometric equations.
}
\details{
This is a helper function for the estimateWoodMass() function.
}
\author{
Courtney L Meier \email{cmeier@battelleecology.org} \cr
}
\keyword{internal}
