% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAllometricWoodyMass.R
\name{estimateAllometricWoodyMass}
\alias{estimateAllometricWoodyMass}
\title{Estimate mass of woody individuals with allometric equations}
\usage{
estimateAllometricWoodyMass(appIndTable, growthFormSubset = NA)
}
\arguments{
\item{appIndTable}{Table (generated in estimateWoodMass()) containing records for qualifying stems of woody individuals. The typical input is a single table created by joining vst_apparentindividual data with the vst_mappingandtagging table (filtered to most recent records per individualID), to obtain 'taxonID', 'scientificName', 'genus', and 'family' fields. \link{data.frame}}

\item{growthFormSubset}{Select Vegetation Structure growth forms for biomass estimation. The options are "tree", which enables biomass estimation only for single- and multi-bole trees with a DBH ≥ 10 cm, and the default of "all", which includes "tree" individuals, and also small trees, single shrubs, small shrubs, and liana individuals. \link{character}}
}
\value{
A table with biomass estimates:
\itemize{
\item For trees, small trees, and lianas: One record for each qualifying bole/stem at each time point it was measured.
\item For shrubs, small shrubs, and saplings: One record for each qualifying individual at each time point it was measured. For multi-stem individuals, input records for multiple stems are combined into a single live and dead biomass estimate per individual.
}
}
\description{
Estimate the biomass of woody individuals via generalized, growth form specific, and species-specific allometric equations.
}
\details{
This is a helper function for the estimateWoodMass() function.
}
\author{
Courtney L Meier \email{cmeier@battelleecology.org} \cr
}
\keyword{internal}
