% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateIncrement.R
\name{estimateIncrement}
\alias{estimateIncrement}
\title{Estimate biomass increment of trees at NEON sites}
\usage{
estimateIncrement(biomassTable, missing, flagged)
}
\arguments{
\item{biomassTable}{Table (generated in estimateWoodProd()) containing biomass records for each tree, with columns for plotID, individualID, taxonID, and its biomass in consecutive years.}

\item{missing}{Select how missing individuals are handled. The default is "filter", which removes trees that were missed during a sampling event and for which plantStatus and biomass cannot be inferred/estimated. These "missing" individuals are filtered to a "missing" table before any productivity calculations are made; in other words, "missing" individuals are ignored for the productivity calculation. The alternate option is "retain", which retains "missing" individuals and assumes that they are dead; this option thus results in missing individuals contributing to both the mortality and increment components of ANPP. \link{character}}

\item{flagged}{Select how individuals flagged for implausibly large stem diameter increments are handled (options are "filter" and "retain"). Default is "retain", which retains flagged individuals and includes them in calculations regardless of stem diameter increment; although flagged individuals contribute to productivity with this option, they are still sent to the "flagged" output table for review. Alternatively, the "filter" option removes all records for individuals with >= 3.5 cm absolute annual stemDiameter increment, including recruited individuals for which inferred initial diameter is 10 cm; flagged individuals are sent to the "flagged" output table. \link{character}}
}
\value{
The input table, with new columns indicating estimated biomass change for each tree in each year.
}
\description{
Calculate the biomass change of trees between time points.
}
\details{
A helper function for the estimateWoodProd() function.
}
\author{
Courtney L Meier \email{cmeier@BattelleEcology.org} \cr
Claire K Lunch \email{clunch@BattelleEcology.org} \cr
}
\keyword{internal}
