% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProd.R
\name{estimateProd}
\alias{estimateProd}
\title{Combine productivity from multiple NEON data products}
\usage{
estimateProd(
  dataProducts = c("Vst", "Hbp"),
  inputWoodMass,
  inputHerbMass = NA,
  plotType = "tower",
  plotPriority = 5,
  calcMethod = "approach_1",
  outlier = 1.5,
  outlierType = "IQR"
)
}
\arguments{
\item{dataProducts}{Specify a character vector with the NEON data products to be combined. Defaults to c("Vst","Hbp") for Vegetation structure and Herbaceous clip harvest, respectively. \link{character}}

\item{inputWoodMass}{Specify a loaded R list object (e.g. estimateWoodMassOutputs) that was produced by companion estimateWoodMass.R function. \link{character}}

\item{inputHerbMass}{Specify a loaded R list object (e.g. scaleHerbMassOutputs) that was produced by companion scaleHerbMass.R function. \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also, if included in the input .rds. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}

\item{calcMethod}{Select plot-level (approach 2) or individual-level (approach 1) productivity calculations (for woody data). The default is "approach_1" \link{character}}

\item{outlier}{Specify how much (if any) outlier removal should be performed (for woody data). The default is 1.5. \link{numeric}}

\item{outlierType}{Specify the type of outlier (for woody data), either SD (standard deviations) or IQR (interquartile range). The default is "IQR". \link{character}}
}
\value{
This function returns a site-level summary dataframe "productivity_site" with NEON productivity (units are Megagrams per hectare per year) from the selected NEON data products.
In addition, list objects (e.g. estimateWoodProdOutputs and estimateHerbProdOutputs) with more granular data summaries are returned for each of the selected NEON data products.
More detailed documentation is provided in the documentation of contributing functions (e.g. estimateWoodProd and estimateHerbProd).
}
\description{
Wrapper function to combine productivity summaries of NEON data products. Currently supported data products are "Vegetation structure"
(DP1.10098.001), estimated with the estimateWoodProd function, and "Herbaceous clip harvest (DP1.10023.001), estimated with the estimateHerbProd function.
The data input is a list of biomass dataframes created by each of the contributing functions. For example, estimateWoodMassOutputs from the estimateWoodMass
function and scaleHerbMassOutputs from the scaleHerbMass function.
}
\details{
A site-level summary of NEON productivity from the selected NEON data products is provided as a dataframe, in addition to list objects with more detailed
outputs for each of the selected NEON data products. For more documentation details see the documentation for the companion functions (e.g. estimateWoodProd
and estimateHerbProd).
}
\examples{
\dontrun{
# example with arguments at default values
estimateProdOutputs <- estimateProd(inputWoodMass = estimateWoodMassOutputs, 
                                    inputHerbMass = scaleHerbMassOutputs)
                      
# example specifying many non-default arguments
estimateWoodProdOutputs <- estimateWoodProd(input = estimateWoodMassOutputs, 
               plotType = "all", plotPriority = 50, 
               calcMethod = "approach_2", outlier = 2, outlierType = "SD")

list2env(estimateWoodProdOutputs ,.GlobalEnv) # unlist all data frames for easier viewing
saveRDS(estimateWoodProdOutputs, 'estimateWoodProdOutputs.rds') # save output locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
