% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWoodMass.R
\name{estimateWoodMass}
\alias{estimateWoodMass}
\title{Estimate above-ground biomass of woody vegetation at NEON sites}
\usage{
estimateWoodMass(
  inputDataList,
  inputIndividual = NA,
  inputMapTag = NA,
  inputNonWoody = NA,
  inputPerPlot = NA,
  plotSubset = "all",
  growthFormSubset = "all"
)
}
\arguments{
\item{inputDataList}{A list object comprised of "Vegetation structure" tables (DP1.10098.001) downloaded using the neonUtilities::loadByProduct() function. Expected input table names are "vst_perplotperyear", "vst_mappingandtagging", and "vst_apparentindividual"; it is optional to include the "vst_non-woody" table in the list. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided for the 'inputIndividual', 'inputMapTag', and 'inputPerPlot' arguments. \link{list}}

\item{inputIndividual}{The 'vst_apparentindividual' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputMapTag}{The 'vst_mappingandtagging' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputNonWoody}{(Optional) The 'vst_non-woody' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputPerPlot}{The 'vst_perplotperyear' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{plotSubset}{Options are the default of "all" (all Tower and Distributed plots), "towerAll" (all plots in the Tower airshed but no Distributed plots), "towerAnnualSubset" (the subset of n=5 Tower plots that are sampled annually), and "distributed" (all Distributed plots, which are sampled at 5-yr intervals and are spatially representative of the NLCD classes at a site). \link{character}}

\item{growthFormSubset}{Select Vegetation Structure growth forms for biomass estimation. The options are "tree", which enables biomass estimation only for single- and multi-bole trees, palm trees, and large tree ferns with a DBH ≥ 10 cm, and the default of "all", which includes "tree" individuals, and also small trees, single shrubs, small shrubs, lianas, small palms, small tree ferns, ocotillo, and xerophyllum individuals. Consult the Vegetation Structure Quick Start Guide and/or the Data Product User Guide for more growth form information. \link{character}}
}
\value{
A list that includes individual-level biomass, plot-level summary biomass, site-level summary biomass, and a table of individuals for which biomass could not be estimated for various reasons. Output tables include:
\itemize{
\item vst_AGB_indiv - Above-ground live and dead standing woody biomass reported for each individual ("kg").
\item vst_AGB_plot - Summary of above-ground total, live, and dead standing woody biomass for each plotID x eventID combination ("Mg/ha"). If argument growthFormSubset == "tree", plots with smaller woody biomass but no trees will have zero biomass in this data frame.
\item vst_AGB_site - Summary of above-ground total, live, and dead standing woody biomass for each siteID x year combination in the data ("Mg/ha").
\item vst_AGB_duplicates - Duplicated individualID x tempStemIDs within a given sampling event (i.e., "eventID"). The individualID x tempStemID should be unique for multi-stem individuals within an eventID; duplicates are identified but not removed before above-ground biomass calculations are carried out.
\item vst_lost_downed - Individuals with 'plantStatus' values of "removed", "lost" of some type, "no longer qualifies", and "downed", and also those individuals for which an allometry was missing or for which an above-ground biomass estimate is missing for some other reason.
\item variables - Units and definitions of novel variables created by the function that are not already defined in the Vegetation Structure data product.
}
}
\description{
Allometric equations are used to estimate above-ground biomass for woody individuals reported in the NEON "Vegetation structure" data product (DP1.10098.001). Input data must be NEON RELEASE-2027 or newer. Results are summarized as mass per unit area at scales of the plotID and siteID. Biomass outputs can be used in the companion estimateWoodProd() function.

Data inputs are either "Vegetation structure" data (DP1.10098.001) in list format retrieved using the neonUtilities::loadByProduct() function (preferred), data tables downloaded from the NEON Data Portal, or input tables with an equivalent structure.
}
\details{
Input data can be filtered via the 'plotSubset' argument if output for only certain types of plots or sampling intervals is desired. Consult the companion getVegStructureEvents() function for a report of which plot types were sampled in which years at a given NEON site. Input data are combined with taxon specific characteristics (e.g., wood density), and biomass is estimated for each individual using allometric equations. Taxon-specific equations are applied if available, and generalized allometries are used otherwise. The 'growthFormSubset' argument enables biomass estimation for "tree" individuals only (i.e., woody individuals with DBH ≥ 10 cm) or for "all" growth forms excluding "cactus", "ferns", and "yucca". Biomass is summarized on an areal basis at the hierarchical levels of the plotID and siteID.
}
\examples{
\dontrun{
#  Obtain NEON Vegetation structure data; note that a token is required and may be obtained after creating a NEON user account
vstDF <- neonUtilities::loadByProduct(
dpID = "DP1.10098.001",
package = "basic",
check.size = FALSE,
token = "my_NEON_token"
)

#  Example with arguments at default values
df <- estimateWoodMass(inputDataList = vstDF)

#  Example specifying non-default arguments for 'plotSubset' and 'growthFormSubset'
df <- estimateWoodMass(
inputDataList = vstDF,
plotSubset = "towerAnnualSubset",
growthFormSubset = "tree"
)

}

}
\author{
Courtney L Meier \email{cmeier@BattelleEcology.org} \cr
Samuel M Simkin \email{samuel.simkin@gmail.com} \cr
}
