% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiomassInputs.R
\name{getBiomassInputs}
\alias{getBiomassInputs}
\title{Download vegetation structure (VST) data and optionally aboveground herbaceous biomass (HBP) data from NEON portal}
\usage{
getBiomassInputs(
  site = "all",
  start = 2018,
  end = as.integer(format(Sys.Date(), "\%Y")) - 1,
  dataProducts = "VstHbp"
)
}
\arguments{
\item{site}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all. \link{list}}

\item{start}{Either NA, meaning all available years starting with 2018, or a character vector specifying start year in the form YYYY, e.g. 2019. The default and earliest allowable option is 2018. \link{character}}

\item{end}{Either NA, meaning all available years after the start year, or a character vector specifying start year in the form YYYY, e.g. 2021. Data from the current calendar year are excluded since they would be incomplete. \link{character}}

\item{dataProducts}{Select whether to analyse only woody biomass "Vst" or woody plus herbaceous "VstHbp" \link{character}}
}
\value{
This function returns a list of lists. The returned parent list includes at least one list, named 'VstDat', with "Vegetation structure" (VST) data from the NEON data portal.
If the dataProducts parameter is "VstHbp" then the returned parent list also includes a second list, named 'HbpDat', with "Herbaceous clip harvest" (HBP) data from the NEON portal.
}
\description{
Download NEON vegetation structure (VST) data and optionally aboveground herbaceous biomass (HBP) data for usage in calculateBiomass function.
}
\details{
All available data from the NEON "Vegetation structure" data product (dpID "DP1.10098.001") meeting the site and year query criteria
will be downloaded using the neonUtilities::loadByProduct function. If dataProducts option "VstHbp" is selected then "Herbaceous clip harvest (dpID "DP1.10023.001")
data are also downloaded using the neonUtilities::loadByProduct function.
}
\examples{
\dontrun{

# example with arguments at default values
VstHbpData <- getBiomassInputs(site="all")

# examples specifying multiple non-default arguments
VstHbpData <- getBiomassInputs(site="HARV", start = 2021, end = 2021, dataProducts = "Vst")

VstHbpData <- getBiomassInputs(siteID = c("HARV","JERC"), 
         start = "2019", end = "2022", dataProducts = "Vst")

list2env(VstHbpData ,.GlobalEnv) # unlist VstDat list (and optionally HbpDat list)
saveRDS(VstHbpData$VstDat, 'VstDat.rds') # save vst data locally
saveRDS(VstHbpData$HbpDat, 'HbpDat.rds') # save hbp data locally

}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
