% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHerbClipSamplingInfo.R
\name{getHerbClipSamplingInfo}
\alias{getHerbClipSamplingInfo}
\title{Find NEON herbaceous clip harvest sampling information}
\usage{
getHerbClipSamplingInfo(sites = NA_character_, token = NA_character_)
}
\arguments{
\item{sites}{4-letter code for one more NEON sites. \link{character}}

\item{token}{NEON API token \link{character}}
}
\value{
A table listing the types of plots sampled for each site-year combination, the plot management activities associated with each plot type, and the number of unique plots and sampling events for each combination. Start and end dates indicate when sampling started in each site-year, and when sampling was completed within each site-year; for sites with more than one sampling event (e.g., grazed sites) this date range encompasses all sampling events. For the most recent year of data, it is possible that not all data collection has been completed; in this case, the end date reflects the end date of data in the database, and the release of more data may be pending.
}
\description{
Herbaceous clip harvest (DP1.10023.001) data are collected annually at NEON terrestrial sites with qualifying vegetation, and different plots are sampled in different years. Moreover, plot management may change through time at a site, with sites in a given year being managed for grazing, agricultural crops, neither, or with distinctly managed subsets of plots within a site. This function returns a table documenting which management activities occurred in which types of plots in a given year for each site.
}
\details{
Herbaceous clip harvest sampling occurs in two types of plots, and management activities within plot type inform how data are used by the scaleHerbMass() and estimateHerbProd() functions:
\itemize{
\item Distributed plots: Sampled once per growing season every 5-6 years with a focus on site-level above-ground biomass; at sites with agricultural crops, these plots planted with crops may be sampled more than one time per year to enable site-level productivity estimates. Grazing exclosures are never deployed in these plots and Distributed plots are not used for site-level herbaceous productivity estimates at grazed sites.
\item Tower plots: Sampled once or multiple times per growing season with a focus on herbaceous productivity. At sites not managed for grazing or agriculture, Tower plots are sampled 1-2 times per year depending on the plant community. At grazed sites, Tower plots subjected to grazing are sampled with and without exclosures, and one to multiple times per year depending on the plant community, to enable estimating grazing consumption and final standing crop at the end of the season; not all Tower plots at grazed sites are grazed. At agricultural sites, each plot is clipped before each crop reaches maturity, meaning each plot may be clipped one or more times per year depending on the site host planting schedule; not all Tower plots at agricultural sites are planted with crops.
}

Additional site-specific information:
\itemize{
\item The San Joaquin Experimental Range (SJER) site: The site experiences a Mediterranean growing season from roughly October to May, and herbaceous plants are largely dormant through the summer months. As such, data from August of one year through July of the following year must be downloaded for this site to capture a full growing season.
}
}
\examples{
\dontrun{

rmnpHerbClipInfo <- getHerbClipSamplingInfo(sites = "RMNP", token = myToken)

herbClipInfo <- getHerbClipSamplingInfo(sites = c("KONZ", "RMNP"), token = myToken)

}

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Courtney Meier \email{cmeier@BattelleEcology.org} \cr
}
