% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVegStructureEvents.R
\name{getVegStructureEvents}
\alias{getVegStructureEvents}
\title{Find NEON Vegetation Structure Sampling Bouts}
\usage{
getVegStructureEvents(
  site = NA_character_,
  includePlots = FALSE,
  token = NA_character_
)
}
\arguments{
\item{site}{4-letter code for a single NEON site. \link{character}}

\item{includePlots}{Include plotIDs for the plots sampled in each event? Defaults to FALSE. \link{logical}}

\item{token}{NEON API token \link{character}}
}
\value{
A table listing the sampling event type for each year and eventID, and the start and end date of each sampling event. If includePlots = TRUE, the table also includes the list of plots sampled in each event. Start and end date are the start and end date of the entire event, not specific to a given plot. For the most recent year of data, it is possible not all data collection has been completed; in this case the end date reflects the end date of data in the database, but publication of more data may be pending.
}
\description{
Vegetation structure (DP1.10098.001) data are collected annually at NEON terrestrial sites with qualifying vegetation, but different plots are sampled in different years. This function returns a table describing which types of sampling events occurred in each year, and the dates of sampling.
}
\details{
Vegetation structure event types are:
\itemize{
\item towerSubset: The subset of plots (n=5) in the Tower airshed that are sampled annually.
\item allTowerPlots: All plots in the Tower airshed, including the annual subset. Tower plots not in the annual subset are sampled every 5-6 years, and always in a different year from the Distributed plots.
\item distributedAndTowerSubset: The annual plots in the Tower airshed plus the Distributed plots. Distributed plots are sampled every 5-6 years.
}

Trees are typically measured outside of the growing season. It is not unusual for sampling bouts to begin in one year and end in the next, so check the start and end date when choosing the date range of data for your analysis.
}
\examples{
\dontrun{

rmnpevents <- getVegStructureEvents(site="RMNP", token=token)

}

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire K Lunch \email{clunch@battelleecology.org} \cr
}
