% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVstHbp.R
\name{getVstHbp}
\alias{getVstHbp}
\title{getVstHbp.R}
\usage{
getVstHbp(
  site = "all",
  start = "2018",
  end = as.character(as.integer(format(Sys.Date(), "\%Y")) - 1),
  dataProducts = "VstHbp"
)
}
\arguments{
\item{site}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all. \link{list}}

\item{start}{Either NA, meaning all available years starting with 2018, or a character vector specifying start year in the form YYYY, e.g. 2019. The default and earliest allowable option is 2018. \link{character}}

\item{end}{Either NA, meaning all available years after the start year, or a character vector specifying start year in the form YYYY, e.g. 2021. Data from the current calendar year are excluded since they would be incomplete. \link{character}}

\item{dataProducts}{Select whether to analyse only woody biomass "Vst" or woody plus herbaceous "VstHbp" \link{character}}
}
\value{
This function returns NEON "Vegetation structure" (VST) data from the NEON data portal, and optionally "Herbaceous clip harvest" (HBP) data as well.
'VstDat' Vegetation structure data from the NEON portal
'HbpDat' Herbaceous clip harvest data from the NEON portal (optional)
}
\description{
Download NEON vegetation structure (VST) data and optionally aboveground herbaceous biomass (HBP) data for usage in NEONbiomass function.

changelog and author contributions / copyrights
Samuel M Simkin (2023-09-14)  original creation
}
\details{
All available data from the NEON "Vegetation structure" data product (dpID "DP1.10098.001") meeting the site and year query criteria
will be downloaded using the neonUtilities::loadByProduct function. If dataProducts option "VstHbp" is selected then "Herbaceous clip harvest (dpID "DP1.10023.001")
data are also downloaded using the neonUtilities::loadByProduct function.
}
\examples{
\dontrun{
# example with arguments at default values
VstHbpData <- getVstHbp(site="all")

# example specifying non-default arguments
VstHbpData <- getVstHbp(siteID = c("HARV","JERC"), 
         start = "2019", end = "2022", dataProducts = "Vst")

list2env(VstHbpData ,.GlobalEnv) # unlist all data frames for easier viewing or additional analysis
saveRDS(VstHbpData$VstDat, 'VstDat.rds') # save vst portal data locally for use in NEONbiomass function
saveRDS(VstHbpData$HbpDat, 'HbpDat.rds') # save hbp portal data locally for use in NEONbiomass function

}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
