% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEONbiomass.R
\name{NEONbiomass}
\alias{NEONbiomass}
\title{NEONbiomass.R}
\usage{
NEONbiomass(
  inputVst = "VstDat.rds",
  inputHbp = "HbpDat.rds",
  site = NA,
  start = NA,
  end = NA,
  growthForm = "single and multi-bole trees",
  plotType = "tower",
  plotPriority = 5
)
}
\arguments{
\item{inputVst}{Specify a local file in .rds format (e.g. VstDat.rds) that contains NEON portal VST data. \link{character}}

\item{inputHbp}{Optionally, specify a local file in .rds format (e.g. vstHbp.rds) that contains NEON portal HBP data. \link{character}}

\item{site}{Either NA, meaning all available sites from inputVst and inputHBP data, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all. \link{list}}

\item{start}{Either NA, meaning all available years from inputVst and inputHBP data, or a character vector specifying start year in the form YYYY, e.g. 2019. The default and earliest allowable option is 2018. \link{character}}

\item{end}{Either NA, meaning all available years from inputVst and inputHBP data after the start year, or a character vector specifying start year in the form YYYY, e.g. 2021. Data from the current calendar year are excluded since they would be incomplete. \link{character}}

\item{growthForm}{Select which growth forms to analyse \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}
}
\value{
This function returns a list that includes biomass summary data frames and a helper data frame
for companion productivity script, and two additional optional data frames if herbaceous data has been analyzed.
'vst_agb_per_ha' summarizes above-ground live woody biomass for each record (units are Megagrams per hectare).
'vst_plot_w_0s' summarizes above-ground live and dead woody biomass for each plot (units are Megagrams per hectare).
'vst_agb_zeros' is a helper for productivity that contains plot x year combos with biomass of 0.
'vst_site' summarizes above-ground live and dead woody biomass for each site (units are Megagrams per hectare).
'hbp_agb_per_ha' summarizes above-ground herbaceous biomass for each record (units are Megagrams per hectare).
'hbp_plot' summarizes above-ground live herbaceous biomass for each plot (units are Megagrams per hectare).
'VstHbp_site' combines above-ground woody and herbaceous biomass for each site.
}
\description{
Use allometric equations to calculate biomass for each VST woody record and
summarise biomass by siteID, plotID, taxonID. Optionally summarise HBP aboveground herbaceous
biomass as well.

changelog and author contributions / copyrights
Samuel M Simkin (2021-03-30)  original creation
Samuel M Simkin (2022-07-12)  revised
Samuel M Simkin (2023-08-04)  revised
Samuel M Simkin (2023-09-14)  revised
}
\details{
All available data from the NEON "Vegetation structure" data product (dpID "DP1.10098.001") meeting the site and year query criteria
will be downloaded using the neonUtilities::loadByProduct function. Supplemental static tables with allometric equation parameters and taxon
specific characteristics are also read in, and biomass is then summarized for each record. Only the set of growth forms selected via the
growthForm parameter are included in the summary. If dataProducts option "VstHbp" is selected then "Herbaceous clip harvest (dpID "DP1.10023.001")
data are also downloaded using the neonUtilities::loadByProduct function and a combined woody and herbaceous summary is provided.
}
\examples{
\dontrun{
# example with arguments at default values
biomassFunctionOutputs <- NEONbiomass(inputVst = "VstDat.rds, inputHbp = "HbpDat.rds, 
        growthForm = "single and multi-bole trees", plotType = "tower", plotPriority = 5)

list2env(NEONbiomassOutputs ,.GlobalEnv) # unlist all data frames for easier viewing or additional analysis
saveRDS(NEONbiomassOutputs, 'NEONbiomassOutputs.rds') # save all outputs locally for further examination and 
       if desired use in the follow-up productivity function.
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
