% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonPlants-package.R
\docType{package}
\name{neonPlants-package}
\alias{neonPlants}
\alias{neonPlants-package}
\title{neonPlants: Tools to access and work with NEON plant data}
\description{
Library of R functions to work with NEON plant data.
}
\author{
\strong{Maintainer}: Eric Sokol \email{esokol@BattelleEcology.org} (\href{https://orcid.org/0000-0001-5923-0917}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eric Sokol \email{esokol@BattelleEcology.org} (\href{https://orcid.org/0000-0001-5923-0917}{ORCID}) [copyright holder]
  \item Dave Barnett \email{dbarnett@BattelleEcology.org} (\href{https://orcid.org/0000-0002-0485-3567}{ORCID})
  \item Sam Simkin \email{ssimkin@BattelleEcology.org} (\href{https://orcid.org/0000-0003-2418-4265}{ORCID})
  \item Courtney Meier \email{cmeier@BattelleEcology.org} (\href{https://orcid.org/0000-0003-3576-883X}{ORCID})
  \item Katherine Jones \email{kjones@BattelleEcology.org} (\href{https://orcid.org/0000-0003-1175-5049}{ORCID})
}

}
\keyword{internal}
