% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEONproductivity.R
\name{NEONproductivity}
\alias{NEONproductivity}
\title{NEONproductivity.R}
\usage{
NEONproductivity(
  input = "NEONbiomassOutputs.rds",
  plotType = "tower",
  plotPriority = 5,
  calcMethod = "approach_1",
  outlier = 1.5,
  outlier_type = "IQR",
  dataProducts = "VstHbp"
)
}
\arguments{
\item{input}{The .rds file with needed biomass input data that was produced by companion NEON_biomass_function.R   Defaults to "NEONbiomassOutputs.rds". \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also, if included in the input .rds. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}

\item{calcMethod}{Select plot-level (approach 2) or individual-level (approach 1) productivity calculations. The default is "approach_1" \link{character}}

\item{outlier}{Specify how much (if any) outlier removal should be performed. The default is 1.5. \link{numeric}}

\item{outlier_type}{Specify the type of outlier, either SD (standard deviations) or IQR (interquartile range). The default is "IQR". \link{character}}

\item{dataProducts}{Select whether to analyse only woody biomass "Vst" or woody plus herbaceous "VstHbp" \link{character}}
}
\value{
This function returns a list that includes productivity summary data frames.
'vst_ANPP_plot_w_taxa' summarises woody ANPP for each year x plot x taxonID combination
'vst_ANPP_plot's woody ANPP for each year x plot combination (taxonIDs are aggregated)
'vst_ANPP_site' summarizes woody ANPP for each site (default is to calculate multi-year mean of each plot and then the mean of all plots at site)
'ANPP_site' adds woody and herbaceous ANPP together (if herbaceous data were in the data input)
'increment_all' contains the increments for each woody individual (approach 1 only)
'increment_outlier' contains the woody individual increments that were removed during outlier removal (approach 1 only)
If calcMethod "approach_2" is selected then only the first four dataframes are returned, and they include an "_2" suffix.
}
\description{
Using inputs from companion NEONbiomass function calculate VST woody yearly increment
and mortality, and thereby annual productivity. Optionally, also calculate HBP productivity and add
to VST for a VST + HBP summary.

changelog and author contributions / copyrights
Samuel M Simkin (2021-03-30)  original creation
Samuel M Simkin (2022-07-12)  revised
Samuel M Simkin (2023-08-04)  revised
}
\details{
Biomass data input files (in .rds format) created by the companion NEON_biomass_function.R script are read in and aboveground net primary
productivity (ANPP) is calculated for woody vegetation (from NEON "Vegetation structure" data product (dpID "DP1.10098.001")). Herbaceous vegetation ANPP
(from NEON "Herbaceous clip harvest (dpID "DP1.10023.001")) is also included if herbaceous data is present in the input file). Both a whole plot
level approach and an individual-level approach calculation method are used, and output from the desired calculation method is returned. The individual-level
approach is currently an underestimate since it includes the growth increment component but not the recruitment component. The recruitment component is
currently very sensitive to stems that were overlooked in previous time periods, but when from true ingrowth can be isolated the recruitment component will be
added. Plots can be filtered to more frequently sampled plots using plotType and plotPriority arguments, and outlier observations can be removed.
}
\examples{
\dontrun{
# example with arguments at default values
NEONproductivityOutputs <- NEONproductivity(input = "NEONbiomassOutputs.rds")
                      
# example specifying many non-default arguments
NEONproductivityOutputs <- NEONproductivity(input = "NEONbiomassOutputs.rds", 
               plotType = "all", plotPriority = 50, 
               calcMethod = "approach_2", outlier = 2, outlier_type = "SD", 
               dataProducts = "Vst")

list2env(NEONbiomassOutputs ,.GlobalEnv) # unlist all data frames for easier viewing or additional analysis
saveRDS(NEONproductivityOutputs, 'NEONproductivityOutputs.rds') # save all outputs locally for further examination
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
