\name{parameters_data}
\docType{data}
\alias{parameters}
\title{Parameters for allometric equations in Chojnacky et al. 2014}
\description{
  This data set gives B0 and B1 parameters for allometric equations for 35 groups, taken from Table 5 of Chojnacky et al. 2014.
}
\usage{parameters}
\format{A dataframe containing allometric parameters for 35 groups.}
\source{journal article}
\references{
  Chojnacky, D. C., Heath, L. S., and Jenkins, J. C. (2014)
  Updated generalized biomass equations for North American tree species.
  Forestry 87:129-151.
}
\keyword{datasets}