% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantIntTrop}
\alias{plantIntTrop}
\title{Dataset Title: Range and nativity status by taxonID}
\format{
A data frame with 9326 rows and 3 variables:
\describe{
\item{taxonID}{Accepted taxonID code associated with binomial latin species epithet}
\item{nativeStatus}{Indicator of native vs introduced status in North America}
\item{tropical}{Indicator of tropical vs temperate range in North America}
}
}
\source{
\url{https://plants.sc.egov.usda.gov/}
}
\usage{
plantIntTrop
}
\description{
USDA Plants temperate vs tropical status and introduced vs native status for woody taxa
}
\keyword{datasets}
