% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priority_plots}
\alias{priority_plots}
\title{Dataset Title: NEON plot sampling priority}
\format{
A data frame with 2005 rows and 3 variables:
\describe{
\item{plotID}{The NEON unique plot identifier}
\item{specificModuleSamplingPriority}{The ranked priority in which NEON plots are sampled}
\item{plotType}{The NEON plot type, either Distributed or Tower}
}
}
\source{
\url{https://github.com/NEONScience/NEON-OS-spatial-data/tree/main/TOS/data}
}
\usage{
priority_plots
}
\description{
The spatial sampling priority for all NEON terrestrial plots
}
\keyword{datasets}
