% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPlantDiv.R
\name{reformatSubplotID}
\alias{reformatSubplotID}
\title{Reformat old subplotIDs}
\usage{
reformatSubplotID(x)
}
\arguments{
\item{x}{old subplotID (text string)}
}
\value{
This function returns a data frame
}
\description{
Helper function to reformat old subplotIDs
}
\author{
Eric Sokol \email{esokol@battelleecology.org} \cr
}
