% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootChemJoin.R
\name{rootChemJoin}
\alias{rootChemJoin}
\title{Join NEON Plant Belowground Biomass Root Mass and Root Chemistry Data Into a Single Table}
\usage{
rootChemJoin(inputMass, inputPool, inputChem)
}
\arguments{
\item{inputMass}{The 'bbc_rootmass' table for the site x month combination(s) of interest.
\link{data.frame}}

\item{inputPool}{The 'bbc_chemistryPooling' table for the site x month combination(s) of
interest. \link{data.frame}}

\item{inputChem}{The 'bbc_rootChemistry' table for the site x month combination(s) of
interest. \link{data.frame}}
}
\value{
A table containing both root mass and root chemistry data in the same row for
different root sizeCategories (i.e., subsampleIDs) within a sampleID. The subsampleIDs in
older data with rootStatus == "dead" do not have root chemistry data, and the function returns
these rows with no chemistry data.
}
\description{
This function uses NEON Plant Belowground Biomass data (DP1.10067.001) retrieved
using the neonUtilities::loadByProduct() function (preferred), downloaded from the NEON Data
Portal, or input data tables with an equivalent structure and representing the same site x month
combinations. The function joins the 'bbc_rootmass', 'bbc_chemistryPooling', and 'bbc_rootChemistry'
tables to generate a single table that contains both mass and chemistry data for each sampleID
and subsampleID (i.e., sizeCategory).
}
\details{
For table joining to be successful, all input data frames must contain data from the
same site x month combination(s). When analytical replicates exist in the 'bbc_rootChemistry'
table, this function returns the mean and concatenates analyte-specific QF values, dataQF
values, and remarks into a single string for all analytical replicates.

In the joined output table, 'rootChemistryDataQF' and 'rootChemistryRemarks' are new fields
that result from the concatenation described above, and there is a new 'analyticalRepCount'
field to document the number of analytical replicates associated with each chemistry data
point.
}
\examples{
\dontrun{
#   Obtain NEON Plant Belowground Biomass data
bbc <- neonUtilities::loadByProduct(
dpID = "DP1.10067.001",
site = "all",
startdate = "2022-07",
enddate = "2022-08",
tabl = "all",
check.size = FALSE
)

#   Join downloaded root data
df <- neonPlants::rootChemJoin(
inputMass = bbc$bbc_rootmass,
inputPool = bbc$bbc_chemistryPooling,
inputChem = bbc$bbc_rootChemistry
)

}

}
\author{
Courtney Meier \email{cmeier@battelleecology.org} \cr
}
