% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootMassScale.R
\name{rootMassScale}
\alias{rootMassScale}
\title{Scale Root Biomass by Size Category to Mass Per Area and Mass Per Volume}
\usage{
rootMassScale(
  inputCore,
  inputMass,
  inputDilution = NA,
  includeFragments = FALSE
)
}
\arguments{
\item{inputCore}{The 'bbc_percore' table for the site x month combination(s) of interest
(required). \link{data.frame}}

\item{inputMass}{The 'bbc_rootmass' table for the site x month combination(s) of interest
(required). \link{data.frame}}

\item{inputDilution}{The 'bbc_dilution' table for the site x month combination(s) of interest
(optional, defaults to NA). \link{data.frame}}

\item{includeFragments}{Indicator to determine whether mass of root fragments < 1 cm length
calculated from dilution sampling should be included in the 'totalDryMass' outputs. Defaults
to FALSE. If set to TRUE, the 'bbc_dilution' table must be provided to the 'inputDilution'
argument. \link{logical}}
}
\value{
A table containing root mass data per unit area ("g/m2") and per unit volume ("g/m3")
for three sizeCategories (< 1mm, 1-2mm, and 2-10mm) as well as total fine root biomass summed
across all sizeCategories. The output no longer contains the 'rootStatus' field.
}
\description{
Join NEON Plant Belowground Biomass data tables (DP1.10067.001) to calculate fine
root biomass by sizeCategory as well as total fine root biomass per unit area and per unit soil
volume. Fine root fragment mass (root fragments < 1 cm length) can optionally be calculated
for the subset of cores for which it is generated, and fragment mass is not included in the summed
total fine root biomass.

Input data should be the 'bbc_percore' and 'bbc_rootmass' tables retrieved using the
neonUtilities::loadByProduct() function (preferred), downloaded from the NEON Data Portal, or
input data tables with an equivalent structure and representing the same site x month
combinations.
}
\details{
NEON weighs a minimum of 5\% of samples a second time so that data users can estimate
the uncertainty associated with different technicians weighing dried roots; QA samples of this
nature are identified via qaDryMass == "Y". The function calculates the mean when QA masses
exist and any 'remarks' are concatenated. Samples with Sampling Impractical values other than "OK"
are removed prior to generating output data.

If inputMass data collected prior to 2019 are provided, the 0-0.5mm and 0.5-1mm sizeCategories
are combined into the current 0-1mm sizeCategory.
}
\examples{
\dontrun{
#   Obtain NEON Plant Belowground Biomass core data
bbc <- neonUtilities::loadByProduct(
dpID = "DP1.10067.001",
site = "all",
startdate = "2022-07",
enddate = "2022-08",
tabl = "all",
check.size = FALSE
)

#   Calculate root mass per unit area and per unit volume
df <- neonPlants::rootMassScale(
inputCore = bbc$bbc_percore,
inputMass = bbc$bbc_rootmass,
inputDilution = bbc$bbc_dilution,
includeFragments = FALSE
)

}

}
\author{
Courtney Meier \email{cmeier@battelleecology.org} \cr
}
