% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootMassStandardize.R
\name{rootMassStandardize}
\alias{rootMassStandardize}
\title{Standardize Root Dry Mass Data to Current Size Category Bins}
\usage{
rootMassStandardize(inputMass)
}
\arguments{
\item{inputMass}{The 'bbc_rootmass' table for the site x month combination(s) of interest
(required). \link{data.frame}}
}
\value{
A table containing root mass data for three sizeCategories (< 1mm, 1-2mm, and 2-10mm)
and dryMass values pooled across previously utilized "live/dead" rootStatus categories. The
output no longer contains the 'rootStatus' field.
}
\description{
NEON root mass data collected prior to 2019 were sorted to four sizeCategories and
were parsed into "live" and "dead" rootStatus categories. Beginning in 2019 the two smallest
sizeCategories (< 0.5mm and 0.5-1mm) were combined into a single < 1mm sizeCategory and roots
were no longer sorted to "live/dead". This function combines the older smallest two sizeCategories
into the current < 1mm category and pools together roots previously categorized as "live" and
"dead". Input data should be the 'bbc_rootmass' table from the NEON Plant Belowground Biomass
data product (DP1.10067.001) retrieved using the neonUtilities::loadByProduct() function
(preferred), downloaded from the NEON Data Portal, or input data tables with an equivalent
structure and representing the same site x month combinations.
}
\details{
NEON weighs a minimum of 5\% of samples a second time so that data users can estimate
the uncertainty associated with different technicians weighing dried roots; QA samples of this
nature are identified via qaDryMass == "Y". The function calculates the mean when QA masses
exist and any 'remarks' are concatenated. Samples with Sampling Impractical values other than "OK"
are removed prior to summarizing the input data.
}
\examples{
\dontrun{
#   Obtain NEON Plant Belowground Biomass data
bbc <- neonUtilities::loadByProduct(
dpID = "DP1.10067.001",
site = "all",
startdate = "2022-07",
enddate = "2022-08",
tabl = "all",
check.size = FALSE
)

#   Standardize downloaded root data to current sizeCategories and rootStatus
df <- neonPlants::rootMassStandardize(
inputMass = bbc$bbc_rootmass
)

}

}
\author{
Courtney Meier \email{cmeier@battelleecology.org} \cr
}
