% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleRootMass.R
\name{scaleRootMass}
\alias{scaleRootMass}
\title{Scale NEON root biomass by size category to mass per area and mass per volume}
\usage{
scaleRootMass(
  inputDataList,
  includeDilution = TRUE,
  inputCore = NA,
  inputMass = NA,
  inputDilution = NA,
  includeFragInTotal = FALSE
)
}
\arguments{
\item{inputDataList}{A list object comprised of Plant Below Ground Biomass tables (DP1.10067.001) downloaded using the neonUtilities::loadByProduct() function. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided for 'inputCore' and 'inputMass' arguments at a minimum. \link{list}}

\item{includeDilution}{Indicator for whether mass of root fragments < 1 cm length should be calculated, as estimated via the Dilution Sampling method (Defaults to TRUE). If TRUE and 'inputDilution' is NA, the 'bbc_dilution' table will be extracted from the list input. \link{logical}}

\item{inputCore}{The 'bbc_percore' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputMass}{The 'bbc_rootmass' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputDilution}{The 'bbc_dilution' table for the site x month combination(s) of interest (optional, defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{includeFragInTotal}{Indicator for whether mass of root fragments < 1 cm length calculated from dilution sampling should be included when summing across sizeCategory to calculate the 'totalDryMass'. Defaults to FALSE. If set to TRUE and 'inputDataList' is missing, the 'bbc_dilution' table must be provided to the 'inputDilution' argument. \link{logical}}
}
\value{
A list containing the following output tables:
\itemize{
\item bbc_BGB_core - Contains root mass data at the scale of the field-collected core, reported for three sizeCategories (< 1mm, 1-2mm, and 2-10mm) as per unit area ("g/m2") and per unit volume ("g/m3"), as well as total fine root biomass summed across all sizeCategories (separate columns for "g/m2", "g/m3", and "Mg/ha"). Output no longer contains the 'rootStatus' field, and QA dryMass samples are averaged.
\item bbc_BGB_plot - Contains mean root mass data at the scale of the plot for each eventID in the data, reported for three sizeCategories (< 1mm, 1-2mm, and 2-10mm) as per unit area ("g/m2") and per unit volume ("g/m3"), as well as total fine root biomass summed across all sizeCategories (separate columns for "g/m2", "Mg/ha", and "g/m3"). Uncertainty at the plot scale is not reported because intra-plot replication is frequently insufficient and not required by the sampling design, but the number of core samples used to calculate the mean is reported.
\item bbc_BGB_site - Provides mean site-level total root mass data for each year (i.e., eventID) in the data, calculated from plot-level data and reported per unit area ("g/m2", "Mg/ha") and per unit volume ("g/m3").
\item variables - Units and definitions of novel variables created by the function that are not already defined in the Plant Belowground Biomass data product.
}
}
\description{
Join NEON Plant Belowground Biomass data tables (DP1.10067.001) to calculate fine root biomass by sizeCategory as well as total fine root biomass per unit area and per unit soil volume. Fine root fragment mass (root fragments < 1 cm length) can optionally be calculated for the subset of cores for which it is generated, and fragment mass can optionally be included in the summed total fine root biomass.

Data inputs are NEON Plant Belowground Biomass data (DP1.10067.001) in list format retrieved using the neonUtilities::loadByProduct() function (preferred), data tables downloaded from the NEON Data Portal, or input data tables with an equivalent structure.
}
\details{
Input data may be provided either as a list or as individual tables. However, if both list and table inputs are provided at the same time the function will error out. For all output data, columns with the same name as input data have identical units and definitions; new columns are described in the 'variables' output table.

If inputMass data collected prior to 2019 are provided, the 0-0.5mm and 0.5-1mm sizeCategories are combined into the current 0-1mm sizeCategory.

NEON weighs a minimum of 5\% of samples a second time so that data users can estimate the uncertainty associated with different technicians weighing dried roots; QA samples of this nature are identified via qaDryMass == "Y". The function calculates the mean when QA masses exist and any 'remarks' are concatenated. Samples with Sampling Impractical values other than "OK" are removed prior to generating output data.
}
\examples{
\dontrun{
#   Obtain NEON Plant Belowground Biomass core data; note that a token is required and may be obtained after creating a NEON user account
bbcDF <- neonUtilities::loadByProduct(
dpID = "DP1.10067.001",
site = "all",
startdate = "2022-07",
enddate = "2022-08",
tabl = "all",
check.size = FALSE,
token = "my_NEON_token"
)

#   Calculate root mass per unit area and per unit volume
df <- neonPlants::scaleRootMass(
inputDataList = bbcDF,
includeDilution = TRUE,
inputCore = NA,
inputMass = NA,
inputDilution = NA,
includeFragInTotal = FALSE
)

}

}
\author{
Courtney Meier \email{cmeier@battelleecology.org} \cr
}
