% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_fields}
\alias{taxon_fields}
\title{Dataset Title: Wood traits for NEON taxonIDs}
\format{
A data frame with 405 rows and 7 variables:
\describe{
\item{taxonID}{Accepted taxonID code associated with binomial latin species epithet}
\item{spg_gcm3}{Wood specific gravity (g/cm3)}
\item{density_source}{Wood specific gravity reference}
\item{decid_vs_ever}{Deciduous vs evergreen growth habit}
\item{decid_vs_ever_source}{Growth habit reference}
\item{woodland_vs_forest}{Typical habitat associated with taxonID}
\item{woodland_vs_forest_source}{Habitat reference}
}
}
\source{
Internal NEON dataset
}
\usage{
taxon_fields
}
\description{
Wood density, deciduous vs evergreen habit, and habitat type for NEON taxonIDs
}
\keyword{datasets}
