\name{taxon_fields_data}
\docType{data}
\alias{taxon_fields}
\title{Taxon-specific wood density and other fields}
\description{
  Some of the Chojnacky et al. 2014 allometric groups require knowledge of taxon-specific wood density and/or whether or not the taxon is deciduous or typically grows in a woodland.
}
\usage{taxon_fields}
\format{A dataframe containing taxonID, spg_gcm3 (wood density in g/cm3), density_source, decid_vs_ever (deciduous vs. evergreen), decid_vs_ever_soruce, woodland_vs_forest, and woodland_vs_forest_source. }
\source{Compiled from various sources. Some values compiled by NEON working group number 75 (WG_75).}
\references{
  NEON
}
\keyword{datasets}