% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables}
\alias{variables}
\title{Dataset Title: Function output variables, units, and definitions}
\format{
A data frame with 215 rows and 5 columns:
\describe{
\item{functionName}{Name of neonPlants function}
\item{table}{Name of function output table}
\item{field}{Name of column in output table}
\item{units}{Column units}
\item{description}{Description of column in output table}
}
}
\source{
Internal NEON dataset
}
\usage{
variables
}
\description{
List of novel variables created by neonPlants functions; does not include variables already defined for existing data products
}
\keyword{datasets}
