% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAquPercentCover.R
\name{estimateAquPercentCover}
\alias{estimateAquPercentCover}
\title{Estimate NEON aquatic plant, bryophyte, lichen, and macroalgae percent cover in wadeable streams}
\usage{
estimateAquPercentCover(
  inputDataList,
  inputPoint = NA,
  inputPerTax = NA,
  inputTaxonomy = NA,
  inputMorph = NA,
  barPlots = FALSE
)
}
\arguments{
\item{inputDataList}{A list object comprised of Aquatic Plant, Bryophyte, Lichen, and Macroalgae Point Count tables (DP1.20072.001) downloaded using the neonUtilities::loadByProduct() function. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided. \link{list}}

\item{inputPoint}{The 'apc_pointTransect' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputPerTax}{The 'apc_perTaxon' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputTaxonomy}{The 'apc_taxonomyProcessed' or 'apc_taxonomyRaw'table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputMorph}{The 'apc_morphospecies' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{barPlots}{If TRUE, will produce a list of plots, one for each site/date in the data provided.}
}
\value{
Percent cover data derived from point counts are returned, along with summary data from sampled transects. In addition, optional graphical output may be returned:
\itemize{
\item apc_percentCover - Contains estimated percent cover for each observed species and/or substrate class on aquatic plant transects.
\item apc_transectMetrics - Contains summary information including the length, habitatType, and total number of points sampled at each transect.
\item apc_barPlots - (Optional) If barPlots = TRUE, a list containing plots for each site x date combination is also produced.
\item variables - Units and definitions of novel variables created by the function that are not already defined in the Aquatic Plant Point Count data product.
}
}
\description{
Data inputs are NEON Aquatic Plant, Bryophyte, Lichen, and Macroalgae Point Counts in Wadeable Streams (DP1.20072.001) in list format retrieved using the neonUtilities::loadByProduct() function (preferred), data tables downloaded from the NEON Data Portal, or input data tables with an equivalent structure and representing the same site x month combinations. The estimateAquPercentCover() function joins taxonomy information across point count tables and aggregates occurrence data to estimate percent cover at the transect level.
}
\details{
Input data may be provided either as a list generated from the neonUtilities::laodByProduct() function or as individual tables. However, only list or table inputs are allowed (not a mix of both).

Percent cover is calculated using the equation from Bowden et al. 2006:

\deqn{percentCover_i = (N_i / N_t) * 100}

Where:
\itemize{
\item \eqn{N_i} is the number of observed points in a transect that match class type “i” (i.e., a particular taxonID or substrate)
\item \eqn{N_t} is the total number of points observed in the transect
}

Note: This calculation can generate percent cover values >100\% if there is vertical stacking of plants, or values <100\% if 'targetTaxaPresent' is unknown.
}
\examples{
\dontrun{
#   Obtain NEON Aquatic Plant Point Count data; note that a token is required and may be obtained after creating a NEON user account
apcDF <- neonUtilities::loadByProduct(
dpID = "DP1.20072.001",
site = "all",
startdate = "2018-07",
enddate = "2018-08",
tabl = "all",
check.size = FALSE,
token = "my_NEON_token"
)

#   Calculate percent cover for downloaded data
df <- neonPlants::estimateAquPercentCover(
inputDataList = apcDF,
inputPoint = NA,
inputPerTax = NA,
inputTaxonomy = NA,
inputMorph = NA,
barPlots = FALSE
)

}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Madaline Ritter \email{ritterm1@battelleecology.org} \cr
}
