% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHerbProd.R
\name{estimateHerbProd}
\alias{estimateHerbProd}
\title{Estimate herbaceous ANPP (Above-ground Net Primary Productivity) at NEON sites}
\usage{
estimateHerbProd(inputDataList, plotSubset = "all")
}
\arguments{
\item{inputDataList}{An R list object produced by the neonUtilities::loadByProduct() function for the NEON Herbaceous Clip Harvest data product. \link{list}}

\item{plotSubset}{The options are the default "all" (all Tower and Distributed plots), "tower" (all plots in the Tower airshed but no Distributed plots), and "distributed" (all Distributed plots, which are sampled on a 5-year interval and are spatially representative of the NLCD classes at a site, and no Tower plots). \link{character}}
}
\value{
A list that includes productivity summary data frames. Output tables include:
\itemize{
\item herb_ANPP_site - Summarizes herbaceous ANPP for each site x year combination ("Mg/ha/yr" and "g/m2/yr").
\item herb_ANPP_plot - Summarizes herbaceous ANPP for the sum of all herbaceous plants for each plot x year combination ("Mg/ha/yr" and "g/m2/yr"). Plot-level summaries are not returned for grazed Tower plots, for the reason outlined in details.
\item herb_ANPP_plot_herbgroup - Summarizes herbaceous ANPP by herbGroup ("Cool Season Graminoids", "N-fixing Plants", etc.) for each plot x year combination ("Mg/ha/yr" and "g/m2/yr"). Plot-level summaries are not returned for grazed Tower plots.
}
}
\description{
Estimate above-ground herbaceous productivity at NEON sites using Herbaceous Clip Harvest data tables (DP1.10023.001).
}
\details{
Input data should be a list of herbaceous biomass (DP1.10023.001) data frames downloaded via the neonUtilities::loadByProduct() function. The companion scaleHerbMass() function is called internally, and outputs from this function are used to estimate herbaceous productivity.

At sites where herbaceous plants are not subject to grazing management and exclosures are not in use, and crops are not planted at any point during the year, the estimate of herbaceous ANPP is simply the standing biomass clipped from the bout with the greatest biomass.

Where grazing management occurs in Tower plots and exclosures have been established, consumption is calculated for each Tower plot sampling event as average biomass within exclosures minus average biomass outside of exclosures. Consumption at each sampling bout is summed and then added to the standing biomass of the last bout of the season for an estimate of herbaceous ANPP.

At sites where crops are planted, productivity is determined on a plot-by-plot basis to account for the potential for multiple plantings throughout the season, as well as "wild-type" plots that are either fallow for the duration of the season or are not managed for agriculture.

Note that at grazed sites, the Science Design only supports calculating consumption at the site level. The clip harvests within exclosures are not close enough to ambient clip harvests to support plot-level consumption estimates.
}
\examples{
\dontrun{
# Obtain NEON Herbaceous clip harvest data
HbpDat <- neonUtilities::loadByProduct(
dpID = "DP1.10023.001",
package = "basic",
check.size = FALSE
)

# Example with arguments at default values
df <-estimateHerbProd(inputDataList = HbpDat)

# Example specifying an alternative plotSubset value
df <-estimateHerbProd(inputDataList = HbpDat,
plotSubset = "tower")

}

}
\author{
Courtney Meier \email{cmeier@BattelleEcology.org} \cr
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
