% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHerbProd.R
\name{estimateHerbProd}
\alias{estimateHerbProd}
\title{Estimate ANPP (Aboveground Net Primary Productivity) contributed by herbaceous vegetation}
\usage{
estimateHerbProd(inputDataList, plotSubset = "towerAnnualSubset")
}
\arguments{
\item{inputDataList}{An R list object produced by the companion scaleHerbMass() function. \link{list}}

\item{plotSubset}{The options are "all" (all tower and distributed plots), the default of "towerAll" (all plots in the tower airshed but no distributed plots), "towerAnnualSubset" (only the subset of tower plots that are sampled annually), and "distributed" (all distributed plots, which are sampled in 5-yr bouts and are spatially representative of the NLCD classes at at site). \link{character}}
}
\value{
A list that includes productivity summary data frames. Output tables include:
\itemize{
\item herb_ANPP_plot - Summarizes herbaceous ANPP for each plot x year combination ("Mg/ha/yr"). Plot level summaries are not returned for grazed sites, for the reason outlined in details.
\item herb_ANPP_site - Summarizes herbaceous ANPP for each site x year combination ("Mg/ha/yr").
}
}
\description{
Calculate herbaceous aboveground productivity using inputs from companion scaleHerbMass function.
}
\details{
A list of herbaceous biomass dataframes created by the companion scaleHerbMass() function is read in, and aboveground net primary productivity (ANPP) is calculated for NEON "Herbaceous clip harvest" (DP1.10023.001) data.

Where herbaceous plants are not subject to grazing management and exclosures are not in use, the estimate of herbaceous ANPP is simply the standing biomass of the last bout of the season.

Where grazing exclosures have been established, consumption is calculated for each sampling event as average biomass within exclosures minus average biomass outside of exclosures. Consumption at each sampling bout is summed and then added to the standing biomass of the last bout of the season for an estimate of herbaceous ANPP.

Note that the Science Design only supports calculating consumption at the site level. The exlosures are not close enough to non-exclosures to support plot level consumption estimates.
}
\examples{
\dontrun{
# Obtain NEON Herbaceous clip harvest data
HbpDat <- neonUtilities::loadByProduct(dpID = "DP1.10023.001",
     package = "basic", check.size = FALSE)

# Use scaleHerbMass to generate output list that is to be used as input to estimateHerbProd
scaleHerbMassOutputs <- neonPlants::scaleHerbMass(
inputDataList = HbpDat,
inputBout = NA,
inputMass = NA
)


# example with arguments at default values
estimateHerbProdOutputs <-estimateHerbProd(inputDataList = scaleHerbMassOutputs)

# example specifying many non-default arguments
estimateHerbProdOutputs <-estimateHerbProd(inputDataList = scaleHerbMassOutputs,
plotSubset = "towerAnnualSubset")

}

}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
