% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHerbProd.R
\name{estimateHerbProd}
\alias{estimateHerbProd}
\title{Estimate NEON plot and site-level ANPP (Aboveground Net Primary Productivity) contributed by herbaceous vegetation}
\usage{
estimateHerbProd(input, plotType = "tower", plotPriority = 5)
}
\arguments{
\item{input}{Specify a loaded R list object (e.g. scaleHerbMassOutputs) that was produced by companion scaleHerbMass function. \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also, if included in the input .rds. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}
}
\value{
This function returns a list that includes productivity summary data frames.
'herb_ANPP_plot' herbaceous ANPP for each plot by year combination (units are Megagrams per hectare per year).
'herb_ANPP_site' herbaceous ANPP for each site by year combination (units are Megagrams per hectare per year).
}
\description{
Using inputs from companion scaleHerbMass function calculate herbaceous aboveground productivity.
The data input is a list of biomass dataframes created by the companion scaleHerbMass function (e.g. scaleHerbMassOutputs).
}
\details{
A list of herbaceous biomass dataframes created by the companion scaleHerbMass function is read in and aboveground net primary
productivity (ANPP) is calculated for NEON "Herbaceous clip harvest" (DP1.10023.001) data. Plots can be filtered to more frequently
sampled plots using plotType and plotPriority arguments. Where grazing exclosures have been established consumption is calculated as
biomass within exclosures minus biomass outside of exclosures. Consumption at each sampling bout is summed and then added to the
standing biomass of the last bout of the season for an estimate of herbaceous ANPP. Where grazing exclosures have not been established
the estimate of herbaceous ANPP is simply the standing biomass of the last bout of the season.
}
\examples{
\dontrun{
# If list is not in memory, load herbaceous biomass list of dataframes from local file:
load('scaleHerbMassOutputs.rds') # load list of dataframes created by scaleHerbMass function
# example with arguments at default values
estimateHerbProdOutputs <-estimateHerbProd(input = scaleHerbMassOutputs)
                      
# example specifying many non-default arguments
estimateHerbProdOutputs <-estimateHerbProd(input = scaleHerbMassOutputs, 
               plotType = "all", plotPriority = 50)

list2env(estimateHerbProdOutputs ,.GlobalEnv) # unlist all data frames for easier viewing
saveRDS(estimateHerbProdOutputs, 'estimateHerbProdOutputs.rds') # save output locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
