% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMass.R
\name{estimateMass}
\alias{estimateMass}
\title{Combine biomass from multiple NEON data products}
\usage{
estimateMass(
  dataProducts = c("Vst", "Hbp"),
  inputDataListVst = NA,
  inputDataListHbp = NA,
  site = NA,
  start = NA,
  end = NA,
  plotType = "tower",
  plotPriority = 5,
  growthForm = "single and multi-bole trees"
)
}
\arguments{
\item{dataProducts}{Specify a character vector with the NEON data products to be combined. Defaults to c("Vst","Hbp") for Vegetation structure and Herbaceous clip harvest, respectively. \link{character}}

\item{inputDataListVst}{Specify a loaded R list object (e.g. VstDat) that contains NEON portal VST (Vegetation structure) data. \link{character}}

\item{inputDataListHbp}{Specify a loaded R list object (e.g. HbpDat) that contains NEON portal HBP (Herbaceous clip harvest) data. \link{character}}

\item{site}{Either NA, meaning all available sites from inputDataListVst and inputHBP data, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all. \link{list}}

\item{start}{Either NA, meaning all available years from inputDataListVst and inputHBP data, or a character vector specifying start year in the form YYYY, e.g. 2019. The default and earliest allowable option is 2018. \link{numeric}}

\item{end}{Either NA, meaning all available years from inputDataListVst and inputHBP data after the start year, or a character vector specifying start year in the form YYYY, e.g. 2021. Data from the current calendar year are excluded since they would be incomplete. [numeric}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}

\item{growthForm}{Select which growth forms to analyse within NEON portal data \link{character}}
}
\value{
This function returns a site-level summary dataframe "biomass_site" with NEON biomass (units are Megagrams per hectare) from the selected NEON data products.
In addition, list objects (e.g. estimateWoodMassOutputs and scaleHerbMassOutputs) with more granular data summaries are returned for each of the selected NEON data products.
More detailed documentation is provided in the documentation of contributing functions (e.g. estimateWoodMass and scaleHerbMass).
}
\description{
Wrapper function to combine biomass summaries of NEON data products. Currently supported data products are "Vegetation structure" (DP1.10098.001),
estimated with the estimateWoodMass function, and "Herbaceous clip harvest (DP1.10023.001), estimated with the scaleHerbMass function. Biomass outputs can, if
desired, be used in the follow-up estimateProd productivity function. Data inputs are list objects retrieved using the neonUtilities::loadByProduct() function
for each selected data product (preferred), or data downloaded from the NEON Data Portal.
}
\details{
A site-level summary of NEON biomass from the selected NEON data products is provided as a dataframe, in addition to list objects with more detailed
outputs for each of the selected NEON data products. For more documentation details see the documentation for the companion functions (e.g. estimateWoodMass
and scaleHerbMass).
}
\examples{
\dontrun{
# If list of lists is not in memory, load lists of dataframes from local files:
load('VstDat.rds') # load NEON VST portal data from a local file
load('HbpDat.rds') # load NEON HBP portal data from a local file

# example with arguments at default values
estimateMassOutputs <- estimateMass(dataProducts = c("Vst","Hbp"), inputDataListVst = VstDat, 
inputDataListHbp = HbpDat )

# example with inputDataListVst = NA that triggers a fresh NEON portal data download
estimateMassOutputs <- estimateMass(inputDataListVst = NA, site = "STEI", 
                                           inputDataListHbp = NA)

# example with just a subset of inputDataListVst data being utilized
estimateMassOutputs <- estimateMass(inputDataListVst = VstDat, start = 2022, 
                                           inputDataListHbp = NA)

# example specifying several non-default arguments
estimateMassOutputs <- estimateMass(inputDataListVst = VstDat, inputDataListHbp = NA,
                                             growthForm = "all trees", plotPriority = 4)

list2env(estimateMassOutputs ,.GlobalEnv) # unlist all data frames
saveRDS(estimateMassOutputs, 'estimateMassOutputs.rds') # save all outputs locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
