% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePheDurationByTag.R
\name{estimatePheDurationByTag}
\alias{estimatePheDurationByTag}
\title{Estimate Phenophase Duration by Tag}
\usage{
estimatePheDurationByTag(
  inputDataList = NULL,
  inputStatus = NULL,
  inputTags = NULL
)
}
\arguments{
\item{pheData}{data list from NEON Plant Phenology Observation (DP1.10055.001) as returned from neonUtilities::loadByProduct(), or individual dataframes for observations and tags as formatted from NEON Plant Phenology Observation (DP1.10055.001) downloaded from neonUtilities::loadByProduct().}
}
\value{
This function returns a data frame
}
\description{
This function uses observation data from the NEON Plant Phenology Observation (DP1.10055.001) to calculate phenophase duration and descriptive statistics.
}
\details{
This function uses the time series created by the neonPlants::estimatePheTransByTag function to calculate phenophase durations  for the time frame provided in the input data frame.
Calculated values inclued:
\itemize{
\item trans_date_start - calendar date of the estimated transition onset
\item trans_doy_start - ordinal day of year of the estimated transition onset
\item trans_date_end - calendar date of the estimated transition end
\item trans_doy_end - ordinal day of year of the estimated transition end
\item duration - difference in days from the onset day of year to the transition end
\item transition_type - indicator that output is for phenophase duration
\item precision_duration - sum of precision_days for estimated oneset and end
\item nth transition - a count of onset events per individualID, phenophase name, within a given calendar year
}
}
\examples{
\dontrun{

# load additional packages for these examples
library(neonUtilities)
library(dplyr)
library(data.table)

# get data
pheDat <- loadByProduct(
  dpID = "DP1.10055.001",
  site = "UKFS",
  startdate = "2022-01",
  enddate = "2022-12",
  package = "basic",
  check.size = FALSE)

out <- estimatePheDurationByTag(inputDataList = pheDat)

WORKS

out2 <- estimatePheDurationByTag(inputStatus = pheDat$phe_statusintensity,
                             inputTags = pheDat$phe_perindividual)

FLAGGING NON EXISTANT DUPLICATES

}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Katie Jones \email{kjones@battelleecology.org} \cr
}
