% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePheTransByTag.R
\name{estimatePheTransByTag}
\alias{estimatePheTransByTag}
\title{Estimate Phenological Transition Dates by Tag}
\usage{
estimatePheTransByTag(
  inputDataList = NULL,
  inputStatus = NULL,
  inputTags = NULL,
  began = FALSE
)
}
\arguments{
\item{inputDataList}{A list of data frames returned from the neonUtilities::loadByProduct() function. \link{list}}

\item{inputStatus}{A data frame with phenological observation data, either the "phe_statusintensity" table or equivalent. \link{data.frame}}

\item{inputTags}{A data frame with taxon data for individuals present in the inputStatus dataframe, either the "phe_perindividual" table or equivalent. \link{data.frame}}

\item{began}{Should transition count be based on the calendar year at the time of transition, or the year the phenophase began? Defaults to FALSE, base the count on calendar year. \link{logical}}
}
\value{
A data frame containing a time series for each phenophase reported for each individual in the data set, including identified  transition dates for beginning and end of a given phenophase, as well as explanatory metrics about that estimate for the time frame provided in the input data frame. Calculated values include:
\itemize{
\item transitionDate - the mid-point between consecutive dates when different phenophase status values were observed.
\item transitionType - onset or end of phenophase.
\item samplingInterval - the number of days between the two observation dates when the phenophase transition occurred.
\item uncertainty - sampling interval/2
\item nthTransition - a count of onset events per individualID and phenophase name within a given year (either calendar year or year phenophase began, depending on the value of the input began=).
}
}
\description{
Data from the NEON Plant Phenology Observation data product (DP1.10055.001) are used to calculate phenophase transition dates for each phenophase transition (status = no -> yes or yes -> no) for each tagged plant or patch observed along a NEON phenology transect or within a phenocam plot in the input data set. Additionally, each estimated transition includes fields describing the count of transitions observed during each year for the given individual x phenophase combination and the sampling interval around the estimate. Required inputs are either a list of data frames (inputDataList) as returned from neonUtilities::loadByProduct() that must include a data frame titled "phe_statusintensity" and one titled "phe_perindividual". Alternatively, the function will accept two individual data frames corresponding to the "phe_statusintensity" table and the "phe_perindividual" table.
}
\details{
Input data may be provided either as a list generated from the neonUtilities::laodByProduct() function or as individual tables. However, if both list and table inputs are provided at the same time the function will error.

For table joining to be successful, inputs must contain data from the same sites for all tables. When individualID duplicates exist in the "phe_perindividual" table, the function will attempt to resolve them based on the editedDate field.

Phenophases may begin in one year and end in another. For the most part this happens in southern sites, but users should be alert for it at any site. Function outputs include both year (the calendar year of the phenophase transition) and yearPhenophaseBegan (the calendar year of phenophase onset for a particular individual and phenophase). Calculation of the count of phenophase transitions per year (nthTransition) can be performed based on either year or yearPhenophaseBegan, using the began= input. The default is year.
}
\examples{
\dontrun{

# load additional packages for these examples
library(neonUtilities)
library(dplyr)
library(data.table)

# get data
pheDat <- loadByProduct(
  dpID = "DP1.10055.001",
  site = "UKFS",
  startdate = "2022-01",
  enddate = "2022-12",
  package = "basic",
  check.size = FALSE)

out <- estimatePheTrans(inputDataList = pheDat)

out2 <- estimatePheTrans(inputStatus = pheDat$phe_statusintensity,
                        inputTags = pheDat$phe_perindividual)
}

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Katie Jones \email{kjones@battelleecology.org} \cr
}
