% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePheTransByTag.R
\name{estimatePheTransByTag}
\alias{estimatePheTransByTag}
\title{Estimate Phenological Transition Dates by Tag}
\usage{
estimatePheTransByTag(
  inputDataList = NULL,
  inputStatus = NULL,
  inputTags = NULL
)
}
\arguments{
\item{pheData}{data list from NEON Plant Phenology Observation (DP1.10055.001) as returned from neonUtilities::loadByProduct().}
}
\value{
This function returns a data frame
}
\description{
This function uses observation data from the NEON Plant Phenology Observation (DP1.10055.001) to calculate phenophase transition dates and descriptive statistics.
}
\details{
This function creates a time series for each phenophase reported for each individual in the data set and identifies transition dates for beginning and end of a given phenophase as well as explanatory metrics about that estimate for the time frame provided in the input data frame.
Calculated values include:
\itemize{
\item transition date - the mid-point between two consecutive dates with different phenophase status values
\item transition_type - indicating the phenophase status values of the transition
\item sampling interval - the number of days between observations
\item uncertainty - sampling interval/2
\item nth transition - a count of onset events per individualID, phenophase name, within a given calendar year
}
}
\examples{
\dontrun{

# load additional packages for these examples
library(neonUtilities)
library(dplyr)
library(data.table)

# get data
pheDat <- loadByProduct(
  dpID = "DP1.10055.001",
  site = "UKFS",
  startdate = "2022-01",
  enddate = "2022-12",
  package = "basic",
  check.size = FALSE)

out <- estimatePheTrans(inputDataList = pheDat)
out2 <- estimatePheTrans(inputStatus = pheDat$phe_statusintensity)
}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Katie Jones \email{kjones@battelleecology.org} \cr
}
