% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateProd.R
\name{estimateProd}
\alias{estimateProd}
\title{Generate summed site-level productivity estimates from NEON data products}
\usage{
estimateProd(
  dataProducts = c("Vst", "Hbp"),
  inputDataListVstMass = NA,
  inputDataListHbpMass = NA,
  plotSubset = "towerAnnualSubset"
)
}
\arguments{
\item{dataProducts}{Specify a character vector with the NEON data products to be combined. Defaults to c("Vst","Hbp") for Vegetation structure and Herbaceous clip harvest, respectively. \link{character}}

\item{inputDataListVstMass}{Specify a loaded R list object (e.g. estimateWoodMassOutputs) that was produced by companion estimateWoodMass or estimateMass function. \link{character}}

\item{inputDataListHbpMass}{Specify a loaded R list object (e.g. scaleHerbMassOutputs) that was produced by companion scaleHerbMass or estimateMass function. \link{character}}

\item{plotSubset}{The options are "all" (all tower and distributed plots), "towerAll" (all plots in the tower airshed but no distributed plots), the default of "towerAnnualSubset" (only the subset of tower plots that are sampled annually), and "distributed" (all distributed plots, which are sampled in 5-yr bouts and are spatially representative of the NLCD classes at at site). \link{character}}
}
\value{
A site-level summary dataframe named "productivity_site" with summed productivity ("Mg/ha/yr") from the selected NEON data products. In addition, more granular data summaries are returned for each of the selected input data products.
}
\description{
Combine above-ground productivity summaries from multiple NEON data products and generate summed site-level productivity estimates. Currently supported data products are "Herbaceous clip harvest" (DP1.10023.001) and "Vegetation structure" (DP1.10098.001).
Data inputs are list objects created by the estimateMass(), scaleHerbMass(), or estimateWoodMass() neonPlants functions.
}
\details{
A site-level summary of NEON productivity from the selected NEON data products is provided as a dataframe, in addition to more detailed outputs for each of the selected NEON data products used to generate the site-level summary. For more documentation details see the documentation for the companion functions - i.e., estimateWoodProd() and estimateHerbProd().
}
\examples{
\dontrun{

#  Example with arguments at default values
estimateProdOutputs <- estimateProd(inputDataListVstMass = estimateWoodMassOutputs,
inputDataListHbpMass = scaleHerbMassOutputs)

#  Example specifying a non-default arguments
estimateWoodProdOutputs <- estimateWoodProd(inputDataListVstMass = estimateWoodMassOutputs,
plotSubset = "all"
)

}

}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
