% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWoodMass.R
\name{estimateWoodMass}
\alias{estimateWoodMass}
\title{Estimate NEON aboveground biomass (AGB) contributed by woody vegetation}
\usage{
estimateWoodMass(
  inputDataListVst,
  site = NA,
  start = NA,
  end = NA,
  growthForm = "single and multi-bole trees",
  plotType = "tower",
  plotPriority = 5
)
}
\arguments{
\item{inputDataListVst}{A list object comprised of "Vegetation structure" tables (DP1.10098.001) downloaded using the neonUtilities::loadByProduct function, or if left at NA then download initiated. \link{character}}

\item{site}{Either NA, meaning all available sites from inputDataListVst data, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all. \link{character}}

\item{start}{Either NA, meaning all available years from inputDataListVst data, or a character vector specifying start year in the form YYYY, e.g. 2019. The default and earliest allowable option is 2018. \link{character}}

\item{end}{Either NA, meaning all available years from inputDataListVst data after the start year, or a character vector specifying start year in the form YYYY, e.g. 2021. Data from the current calendar year are excluded since they would be incomplete. \link{character}}

\item{growthForm}{Select which growth forms to analyse \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}
}
\value{
This function returns a list that includes biomass summary data frames and a helper data frame
for companion productivity script.
'vst_agb_per_ha' summarizes above-ground live woody biomass for each individual (units are Megagrams per hectare).
'vst_plot_w_0s' summarizes above-ground live and dead woody biomass for each plot (units are Megagrams per hectare).
'vst_agb_zeros' is a helper for productivity that contains plot by year combos with biomass of 0.
'vst_site' summarizes above-ground live and dead woody biomass for each site (units are Megagrams per hectare).
}
\description{
Use allometric equations to calculate biomass for each VST woody record and summarise biomass by siteID, plotID, and taxonID.
Biomass outputs can, if desired, be used in the follow-up estimateWoodProd productivity function. Data inputs are "Vegetation structure"
(DP1.10098.001) retrieved using the neonUtilities::loadByProduct() function (preferred), or data downloaded from the NEON Data Portal.
}
\details{
Input data can be filtered by site, date, plot type, and plot priority. Input data are combined with supplemental static tables
with allometric equation parameters and taxon specific characteristics, and biomass is estimated by allometric equations for each individual.
Generalized allometric equations are applied first, and replaced by taxon specific equations if available. Only the set of growth forms selected via the
growthForm parameter are included in outputs. Biomass is summarized on an areal basis at the hierarchical level of the plot and site.
}
\examples{
\dontrun{
# Obtain NEON Vegetation structure
VstDat <- neonUtilities::loadByProduct(dpID="DP1.10098.001", 
     site = "all",
     startdate = "2018-01",
     enddate = paste0(as.character(as.integer(format(Sys.Date(), "\%Y"))-1), "-12"),
     package = "basic", check.size = FALSE)
     
# If list is not in memory, load VST list of dataframes from local file:
load('VstDat.rds') # load NEON VST portal data from a local file

# example with arguments at default values
estimateWoodMassOutputs <- estimateWoodMass(inputDataListVst = VstDat)

# example with inputDataListVst = NA that triggers a fresh NEON portal data download
estimateWoodMassOutputs <- estimateWoodMass(inputDataListVst = NA, site = "STEI")

# example with just a subset of inputDataListVst data being utilized
estimateWoodMassOutputs <- estimateWoodMass(inputDataListVst = VstDat, start = 2022)

# example specifying several non-default arguments
estimateWoodMassOutputs <- estimateWoodMass(inputDataListVst = VstDat,
                                             growthForm = "all trees", plotPriority = 4)

list2env(estimateWoodMassOutputs ,.GlobalEnv) # unlist all data frames
saveRDS(estimateWoodMassOutputs, 'estimateWoodMassOutputs.rds') # save all outputs locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
