% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWoodProd.R
\name{estimateWoodProd}
\alias{estimateWoodProd}
\title{Estimate woody ANPP for a NEON site}
\usage{
estimateWoodProd(
  inputDataList,
  inputIndividual = NA,
  inputMapTag = NA,
  inputPerPlot = NA,
  inputNonWoody = NA,
  plotSubset = "all",
  flagged = "retain",
  missing = "filter"
)
}
\arguments{
\item{inputDataList}{A list object comprised of "Vegetation structure" tables (DP1.10098.001) for a single site, downloaded using the neonUtilities::loadByProduct() function. Expected input table names are "vst_perplotperyear", "vst_mappingandtagging", and "vst_apparentindividual"; it is optional to include the "vst_non-woody" table in the list. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided for the 'inputIndividual', 'inputMapTag', and 'inputPerPlot' arguments. \link{list}}

\item{inputIndividual}{The 'vst_apparentindividual' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputMapTag}{The 'vst_mappingandtagging' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputPerPlot}{The 'vst_perplotperyear' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputNonWoody}{The 'vst_non-woody' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{plotSubset}{Options are the default of "all" (all Tower and Distributed plots), "towerAll" (all plots in the Tower airshed but no Distributed plots), "towerAnnualSubset" (the subset of n=5 Tower plots that are sampled annually), and "distributed" (all Distributed plots, which are sampled at 5-yr intervals and are spatially representative of the NLCD classes at a site). \link{character}}

\item{flagged}{Select how to handle individuals flagged for implausibly large stem diameter increments: "retain" (default) includes flagged individuals in calculations and outputs them to a "flagged" table for review, or "filter" removes all records with ≥3.5 cm absolute annual stem diameter increment (including recruits with inferred initial diameter ≥10 cm) and sends flagged individuals to the "flagged" table. \link{character}}

\item{missing}{Select how to handle individuals missed during sampling and for which plantStatus and biomass cannot be inferred/estimated: "filter" (default) removes missed individuals before productivity calculations and collates them in a "missing" table. The "retain" option assumes missing individuals are dead, and these individuals may contribute to ANPP. \link{character}}
}
\value{
A list that includes productivity summary data frames. Output tables are:
\itemize{
\item vst_ANPP_indiv - Woody ANPP for each individual at each time step for which data exist ("Mg/ha/yr").
\item vst_ANPP_plot - Summarizes woody ANPP for each plot x year combination ("Mg/ha/yr").
\item vst_ANPP_site - Summarizes woody ANPP for each site x year combination ("Mg/ha/yr").
\item vst_ANPP_duplicates - Duplicated single- or multi-bole tree individualIDs within a given sampling event (i.e., "eventID"). The individualID should be unique for single- and multi-bole tree growthForms within an eventID; duplicates are removed before productivity calculations are carried out.
\item vst_ANPP_flagged - Individuals flagged for changes in stemDiameter > 3.5 cm/yr; includes records from all time points for flagged individuals. By default, the 'flagged' argument is "retain" and the records in this table are included in the productivity calculation.
\item vst_ANPP_missing - Individuals that were missed during a sampling event; table is populated only when the 'missing' argument is set to "filter" (default).
\item variables - Units and definitions of novel variables created by the function that are not already defined in the Vegetation Structure data product.
}
}
\description{
Calculate above-ground net primary productivity of trees reported in the NEON "Vegetation structure" data product (DP1.10098.001). Data must be provided one site at a time. Results are summarized as mass per unit area per year at scales of the plotID and siteID.

Data inputs are "Vegetation structure" data for a single site (DP1.10098.001) in list format, either provided via the neonUtilities::loadByProduct() function (preferred), as data tables downloaded from the NEON Data Portal, or as input tables with an equivalent structure and representing the same site x month combinations.

Data must be provided to the function one site at a time, and the 'vst_mappingandtagging' table should include all years of data from the beginning of collection to the last year being analyzed. Returning the full dataset in 'vst_mappingandtagging' is the default behavior of neonUtilities::loadByProduct().
}
\details{
The input data are passed to the companion estimateWoodMass() function to estimate biomass for qualifying trees, and then aboveground net primary productivity is calculated for live trees at each timepoint. Input data are filtered by the 'plotSubset' argument if output for only certain types of plots or sampling intervals is desired. Productivity is summarized on an areal basis with units "Mg/ha/yr" at the hierarchical level of the plot and site.

For trees, the individual-level approach to calculating productivity is used from Clark DA, S Brown, DW Kicklighter, JQ Chambers, JR Thomlinson, and J Ni. 2001. Measuring Net Primary Production in Forests: Concepts and Field Methods. Ecological Applications 11:356-370. With this approach, NEON data enable calculating woody productivity for individuals with a growthForm of "single bole tree" or "multi-bole tree".

NEON has an extensive data QA/QC process, but users should be aware that productivity estimates are very sensitive to data entry errors and so the function output should be examined carefully.
}
\examples{
\dontrun{
# Obtain NEON Vegetation structure for a single site; note that a token is required and may be obtained after creating a NEON user account
vstDF <- neonUtilities::loadByProduct(
dpID = "DP1.10098.001",
site = "ABBY",
package = "basic",
check.size = FALSE,
token = "my_NEON_token"
)

woodProdOutput <- neonPlants::estimateWoodProd(inputDataList = vstDF)

}

}
\author{
Courtney L Meier \email{cmeier@BattelleEcology.org} \cr
Claire K Lunch \email{clunch@BattelleEcology.org} \cr
}
