% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWoodProd.R
\name{estimateWoodProd}
\alias{estimateWoodProd}
\title{Estimate NEON plot and site-level ANPP (Aboveground Net Primary Productivity) contributed by woody vegetation}
\usage{
estimateWoodProd(
  input,
  plotType = "tower",
  plotPriority = 5,
  calcMethod = "approach_1",
  outlier = 1.5,
  outlierType = "IQR"
)
}
\arguments{
\item{input}{Specify a loaded R list object (e.g. estimateWoodMassOutputs) that was produced by companion estimateWoodMass function. \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Defaults to "tower" plots, which are sampled annually. Otherwise "distributed" plots are examined also, if included in the input .rds. \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is 5. \link{numeric}}

\item{calcMethod}{Select plot-level (approach 2) or individual-level (approach 1) productivity calculations. The default is "approach_1" \link{character}}

\item{outlier}{Specify how much (if any) outlier removal should be performed. The default is 1.5. \link{numeric}}

\item{outlierType}{Specify the type of outlier, either SD (standard deviations) or IQR (interquartile range). The default is "IQR". \link{character}}
}
\value{
This function returns a list that includes productivity summary data frames.
'vst_ANPP_plot_w_taxa' summarises woody ANPP for each plot by year by taxonID combination (units are Megagrams per hectare per year).
'vst_ANPP_plot' summarises woody ANPP for each plot by year combination (units are Megagrams per hectare per year).
'vst_ANPP_site' summarizes woody ANPP for each site by year combination (units are Megagrams per hectare per year).
'increment_all' contains the increments for each woody individual (approach 1 only).
'increment_outlier' contains the woody individual increments that were removed during outlier removal (approach 1 only).

If calcMethod "approach_2" is selected then only the first three dataframes are returned, and they include an "_2" suffix.
}
\description{
Using inputs from companion estimateWoodMass function calculate annual woody productivity.
The data input is a list of biomass dataframes created by the companion estimateWoodMass function (e.g. estimateWoodMassOutputs).
}
\details{
A list of woody biomass dataframes created by the companion estimateWoodMass function is read in and aboveground net primary
productivity (ANPP) is calculated for woody vegetation from NEON "Vegetation structure" (DP1.10098.001) data. Both a whole plot level
approach and an individual-level approach calculation method are used, and output from the desired calculation method is returned. The
individual-level approach is currently an underestimate since it includes the growth increment component but not the recruitment component.
The recruitment component is currently very sensitive to stems that were overlooked in previous time periods, but when from true ingrowth
can be isolated the recruitment component will be added. Plots can be filtered to more frequently sampled plots using plotType and
plotPriority arguments, and outlier observations can be removed.
}
\examples{
\dontrun{
# If list is not in memory, load woody biomass list of dataframes from local file:
load('estimateWoodMassOutputs.rds') # load list of dataframes created by estimateWoodMass function
# example with arguments at default values
estimateWoodProdOutputs <- estimateWoodProd(input = estimateWoodMassOutputs)
                      
# example specifying many non-default arguments
estimateWoodProdOutputs <- estimateWoodProd(input = estimateWoodMassOutputs, 
               plotType = "all", plotPriority = 50, 
               calcMethod = "approach_2", outlier = 2, outlierType = "SD")

list2env(estimateWoodProdOutputs ,.GlobalEnv) # unlist all data frames for easier viewing
saveRDS(estimateWoodProdOutputs, 'estimateWoodProdOutputs.rds') # save output locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
