% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWoodProd.R
\name{estimateWoodProd}
\alias{estimateWoodProd}
\title{Estimate ANPP (Aboveground Net Primary Productivity) contributed by woody vegetation}
\usage{
estimateWoodProd(inputDataList, siteID, plotSubset = "towerAnnualSubset")
}
\arguments{
\item{inputDataList}{Specify a loaded R list object (e.g. estimateWoodMassOutputs) that was produced by the companion estimateWoodMass function (and subsequently filtered to two eventIDs if needed). \link{character}}

\item{siteID}{Specify the four-letter siteID for a single NEON site (only one siteID is allowed). \link{character}}

\item{plotSubset}{The available options for this function are the default of "towerAnnualSubset" (only the subset of tower plots that are sampled annually) or "towerAll" (all plots in the tower airshed but no distributed plots). \link{character}
#'}
}
\value{
A list that includes productivity summary data frames. Output tables include:
\itemize{
\item vst_ANPP_plot - Summarizes woody ANPP for each plot x year combination ("Mg/ha/yr").
\item vst_ANPP_site - Summarizes woody ANPP for each site x year combination ("Mg/ha/yr").
}
}
\description{
Calculate annual productivity of woody vegetation (and "non-woody perennial" vegetation, if included in inputDataList) using inputs from the companion estimateWoodMass() function.
}
\details{
An input data list of woody biomass dataframes needs to be created by the companion estimateWoodMass() function and filtered to two eventIDs if needed. This input list object is read in and aboveground net primary productivity (ANPP) is calculated for woody vegetation from NEON "Vegetation structure" (DP1.10098.001) data.
The stand-level approach (approach 2) is used from Clark DA, S Brown, DW Kicklighter, JQ Chambers, JR Thomlinson, and J Ni. 2001. Measuring Net Primary Production in Forests: Concepts and Field Methods. Ecological Applications 11:356-370.
Woody productivity is only calculated for trees with growthForm of "single bole tree" or "multi-bole tree". Therefore, in estimateWoodMass function, specify growthForm "tree".
NEON has an extensive data QA/QC process, but users should be aware that these productivity estimates are very sensitive to any residual errors and so the data should be examined carefully
}
\examples{
\dontrun{
# Obtain NEON Vegetation structure
VstDat <- neonUtilities::loadByProduct(
dpID="DP1.10098.001",
package = "basic",
check.size = FALSE
)

# Use estimateWoodMass to generate output list that is to be used as input to estimateWoodProd
estimateWoodMassOutputs <- estimateWoodMass(inputDataList = VstDat)


# example with arguments at default values
estimateWoodProdOutputs <- estimateWoodProd(inputDataList = estimateWoodMassOutputs)


}

}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
