% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVegStructureEvents.R
\name{getVegStructureEvents}
\alias{getVegStructureEvents}
\title{Find vegetation structure sampling bouts}
\usage{
getVegStructureEvents(
  site = NA_character_,
  includePlots = FALSE,
  token = NA_character_
)
}
\arguments{
\item{site}{4-letter code for a single NEON site. \link{character}}

\item{includePlots}{Include plotIDs for the plots sampled in each event? Defaults to FALSE. \link{logical}}

\item{token}{NEON API token \link{character}}
}
\value{
A table listing the sampling event type for each year and eventID, and the start and end date of each sampling event. If includePlots=T, the table also includes the list of plots measured in each event. Start and end date are the start and end date of the entire event, not specific to a given plot. For the most recent year of data, it is possible not all data collection has been completed; in this case the end date reflects the end date of data in the database, but more may be pending.
}
\description{
Vegetation structure (DP1.10098.001) data are collected annually at NEON terrestrial sites with qualifying vegetation, but different plots are sampled in different years. This function returns a table of which types of sampling events occurred in each year.
}
\details{
Vegetation structure event types are:
towerSubset: The subset of plots in the tower airshed that are measured annually.
allTowerPlots: All plots in the tower airshed, including the annual subset. Tower plots not in the annual subset are measured every 5-6 years, in a different year from the distributed plots.
distributedAndTowerSubset: The annual plots in the tower airshed plus the distributed plots. Distributed plots are measured every 5-6 years.

Trees are typically measured outside of the growing season. It is not unusual for sampling bouts to begin in one year and end in the next, so check the start and end date when choosing the date range of data for your analysis.
}
\examples{
\dontrun{

rmnpevents <- getVegStructureEvents(site="RMNP", token=token)

}

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Claire Lunch \email{clunch@battelleecology.org} \cr
}
