% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAquPointCount.R
\name{joinAquPointCount}
\alias{joinAquPointCount}
\title{Join NEON aquatic plant point count data into a single table with merged taxonomic identifications}
\usage{
joinAquPointCount(
  inputDataList,
  inputPoint = NA,
  inputPerTax = NA,
  inputTaxProc = NA,
  inputMorph = NA
)
}
\arguments{
\item{inputDataList}{A list object comprised of Aquatic Plant, Bryophyte, Lichen, and Macroalgae Point Count tables (DP1.20072.001) downloaded using the neonUtilities::loadByProduct() function. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided. \link{list}}

\item{inputPoint}{The 'apc_pointTransect' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputPerTax}{The 'apc_perTaxon' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputTaxProc}{The 'apc_taxonomyProcessed' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputMorph}{The 'apc_morphospecies' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}
}
\value{
A table containing point transect data with all associated taxonomic information for each point where targetTaxaPresent == 'Y'.
}
\description{
Join the 'apc_pointTransect', 'apc_perTaxon', 'apc_taxonomyProcessed' and 'apc_morphospecies' tables to generate a single table that contains point count data with taxonomic identifications for each sampleID. Data inputs are NEON Aquatic Plant, Bryophyte, Lichen, and Macroalgae Point Counts in Wadeable Streams (DP1.20072.001) in list format retrieved using the neonUtilities::loadByProduct() function (preferred), data tables downloaded from the NEON Data Portal, or input data tables with an equivalent structure and representing the same site x month combinations.
}
\details{
Input data may be provided either as a list or as individual tables. However, if both list and table inputs are provided at the same time the function will error out. For table joining to be successful, inputs must contain data from the same site x month combination(s) for all tables.

In the joined output table, the 'acceptedTaxonID' and associated taxonomic fields are populated from the first available identification in the following order: 'apc_taxonomyProcessed', 'apc_perTaxon', or 'apc_morphospecies'. For samples identified both in the field and by an expert taxonomist, the expert identification is retained in the output. A new field, 'taxonIDSourceTable', is included in the output and indicates the source table for each sample's identification.

If a single sample in 'apc_taxonomyProcessed' contains multiple macroalgae species, each species will be represented as a separate row in 'apc_pointTransect' for every point associated with that sampleID.
}
\examples{
\dontrun{
#   Obtain NEON Aquatic Plant Point Count data
apc <- neonUtilities::loadByProduct(
dpID = "DP1.20072.001",
site = "all",
startdate = "2018-07",
enddate = "2018-08",
tabl = "all",
check.size = FALSE
)

#   Join downloaded root data
df <- neonPlants::joinAquPointCount(
inputDataList = apc,
inputPoint = NA,
inputPerTax = NA,
inputTaxProc = NA,
inputMorph = NA
)

}

}
\author{
Madaline Ritter \email{ritterm1@battelleecology.org} \cr
}
