% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonPlants-package.R
\docType{package}
\name{neonPlants-package}
\alias{neonPlants}
\alias{neonPlants-package}
\title{neonPlants: Tools to access and work with NEON plant data}
\description{
Library of R functions to work with NEON plant data.
}
\author{
\strong{Maintainer}: Courtney Meier \email{cmeier@BattelleEcology.org} (\href{https://orcid.org/0000-0003-3576-883X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Courtney Meier \email{cmeier@BattelleEcology.org} (\href{https://orcid.org/0000-0003-3576-883X}{ORCID}) [copyright holder]
  \item Dave Barnett \email{dbarnett@BattelleEcology.org} (\href{https://orcid.org/0000-0002-0485-3567}{ORCID})
  \item Sam Simkin \email{samuel.simkin@gmail.com} (\href{https://orcid.org/0000-0003-2418-4265}{ORCID})
  \item Katherine Jones \email{joneseyk@gmail.com} (\href{https://orcid.org/0000-0003-1175-5049}{ORCID})
  \item Claire Lunch \email{clunch@BattelleEcology.org} (\href{https://orcid.org/0000-0001-8753-6593}{ORCID})
}

}
\keyword{internal}
