% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonPlants-package.R
\docType{package}
\name{neonPlants-package}
\alias{neonPlants}
\alias{neonPlants-package}
\title{neonPlants: Tools to access and work with NEON plant data}
\description{
Library of R functions to work with NEON plant data.
}
\author{
\strong{Maintainer}: Eric Sokol \email{esokol@battelleecology.org} (\href{https://orcid.org/0000-0001-5923-0917}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Dave Barnett
  \item Sam Simkin \email{ssimkin@battelleecology.org} (\href{https://orcid.org/0000-0003-2418-4265}{ORCID})
}

}
\keyword{internal}
