% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Dataset Title: Chojnacky parameters}
\format{
A data frame with 35 rows and 10 variables:
\describe{
\item{allometry_ID}{Allometry ID name developed in Chojnacky et al. 2014}
\item{Group}{High-level group for which allometry parameters are applicable: either conifer, hardwood, or woodland}
\item{subgroup}{Sub-group for which allometry parameters are applicable - e.g., family, wood density range, etc."}
\item{Median.specific.gravity}{specific gravity of wood on green volume to dry-weight basis}
\item{b0}{Chojnacky parameter b0}
\item{b1}{Chojnacky parameter b1}
\item{Diameter}{Type of diameter required for allometric equation, either dbh or drc}
\item{minDiameter}{Minimum diameter for which parameters are applicable}
\item{maxDiameter}{Maximum diameter for which parameters are applicable}
\item{R2.statistic}{R2 fit for Chojnacky equation}
}
}
\source{
\url{https://doi.org/10.1093/forestry/cpt053}
}
\usage{
parameters
}
\description{
Allometric parameters for 35 different woody biomass equations (from Chojnacky et al. 2014, Table 5)
}
\keyword{datasets}
