% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleHerbMass.R
\name{scaleHerbMass}
\alias{scaleHerbMass}
\title{Scale herbaceous biomass by functional group data to mass per area}
\usage{
scaleHerbMass(
  inputDataList,
  inputBout = NA,
  inputMass = NA,
  plotSubset = "all"
)
}
\arguments{
\item{inputDataList}{A list object comprised of "Herbaceous clip harvest" tables (DP1.10023.001) downloaded using the neonUtilities::loadByProduct() function. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided for 'inputBout', and 'inputMass' arguments. \link{list}}

\item{inputBout}{The 'hbp_perbout' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{inputMass}{The 'hbp_massdata' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{plotSubset}{The options are the default of "all" (all Tower and Distributed plots), "tower" (all plots in the Tower airshed but no Distributed plots), and "distributed" (all Distributed plots, which are sampled on a 5-year interval and are spatially representative of the NLCD classes at a site, and no Tower plots). \link{character}}
}
\value{
A list that includes biomass summary data at multiple scales. Output tables include:
\itemize{
\item hbp_agb - Summarizes above-ground herbaceous biomass for each sampleID in the input data ("g/m2").
\item hbp_plot - Summarizes above-ground herbaceous biomass for each plot by year combination (both "g/m2" and "Mg/ha").
\item hbp_site - Summarizes above-ground herbaceous biomass for each site by year combination (both "g/m2" and "Mg/ha").
}
}
\description{
Join NEON Herbaceous Clip Harvest data tables (DP1.10023.001) to calculate herbaceous biomass by functional group per unit area as well as total herbaceous biomass per unit area. Biomass outputs can be used in the neonPlants estimateMass() function, and the estimateHerbProd() productivity function.

Data inputs are "Herbaceous clip harvest" data (DP1.10023.001) in list format retrieved using the neonUtilities::loadByProduct() function (preferred), data tables downloaded from the NEON Data Portal, or input tables with an equivalent structure and representing the same site x month combinations.
}
\details{
Input data can be filtered by plot subset. Herbaceous biomass data are scaled to an area basis at the hierarchical levels of sampling cell, plot, and site. Input data may be provided either as a list or as individual tables. However, if both list and table inputs are provided at the same time the function will error out. For all output data, columns with the same name as input data have identical units and definitions; where needed, new columns contain new units information.

NEON weighs a minimum of 5\% of samples a second time so that data users can estimate the uncertainty associated with different technicians weighing dried herbaceous biomass; QA samples of this nature are identified via qaDryMass == "Y". The function calculates the mean when QA masses exist. Samples with Sampling Impractical values other than "OK" are removed prior to generating output data.
}
\examples{
\dontrun{
# Obtain NEON Herbaceous clip harvest data
HbpDat <- neonUtilities::loadByProduct(dpID = "DP1.10023.001",
     package = "basic", check.size = FALSE)

# example with arguments at default values
df <- neonPlants::scaleHerbMass(
inputDataList = HbpDat,
inputBout = NA,
inputMass = NA
)

}

}
\author{
Courtney Meier \email{cmeier@BattelleEcology.org} \cr
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
