% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleHerbMass.R
\name{scaleHerbMass}
\alias{scaleHerbMass}
\title{Scale NEON aboveground biomass (AGB) contributed by herbaceous vegetation}
\usage{
scaleHerbMass(
  inputDataList,
  hbp_perbout = NA,
  hbp_massdata = NA,
  plotType = "all",
  plotPriority = NA
)
}
\arguments{
\item{inputDataList}{A list object comprised of "Herbaceous clip harvest" tables (DP1.10023.001) downloaded using the neonUtilities::loadByProduct function.
If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided for
'hbp_perbout', and 'hbp_massdata' arguments. \link{list}}

\item{hbp_perbout}{The 'hbp_perbout' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{hbp_massdata}{The 'hbp_massdata' table for the site x month combination(s) of interest
(defaults to NA). If table input is provided, the 'inputDataList' argument must be missing. \link{data.frame}}

\item{plotType}{Optional filter based on NEON plot type. Options are "all" (both tower and distributed plot types) or "tower". Defaults to "all". \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is NA. \link{numeric}}
}
\value{
This function returns a list that includes biomass summary data frames.
'hbp_agb' summarizes above-ground herbaceous biomass for each record (units are g per m2).
'hbp_plot' summarizes above-ground live herbaceous biomass for each plot by year combination (units are both g per m2 and Megagrams per hectare).
'hbp_site' summarizes above-ground live herbaceous biomass for each site by year combination (units are both g per m2 and Megagrams per hectare).
}
\description{
Summarise aboveground herbaceous biomass. Biomass outputs can, if desired, be used in the follow-up estimateHerbProd productivity function.
Data inputs are "Herbaceous clip harvest" (DP1.10023.001) retrieved using the neonUtilities::loadByProduct() function (preferred), or data downloaded from the NEON Data Portal.
}
\details{
Input data can be filtered by site, date, plot type, and plot priority. Herbaceous biomass are scaled to an area basis at the hierarchical levels of clip cell, plot, and site.
}
\examples{
\dontrun{
# Obtain NEON Herbaceous clip harvest data
HbpDat <- neonUtilities::loadByProduct(dpID = "DP1.10023.001",
     package = "basic", check.size = FALSE)
     
# If list is not in memory, load HBP list of dataframes from local file:
load('HbpDat.rds') # load NEON HBP portal data from a local file

# example with arguments at default values
scaleHerbMassOutputs <- scaleHerbMass(inputDataList = HbpDat, hbp_perbout = NA, hbp_massdata = NA)

list2env(scaleHerbMassOutputs ,.GlobalEnv) # unlist all data frames
saveRDS(scaleHerbMassOutputs, 'scaleHerbMassOutputs.rds') # save all outputs locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
