% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleHerbMass.R
\name{scaleHerbMass}
\alias{scaleHerbMass}
\title{Scale NEON aboveground biomass (AGB) contributed by herbaceous vegetation}
\usage{
scaleHerbMass(
  inputDataListHbp,
  site = NA,
  start = NA,
  end = NA,
  plotType = "all",
  plotPriority = NA
)
}
\arguments{
\item{inputDataListHbp}{A list object comprised of "Herbaceous clip harvest" tables (DP1.10023.001) downloaded using the neonUtilities::loadByProduct function, or if left at NA then download initiated. \link{character}}

\item{site}{Either NA, meaning all available sites from inputHBP data, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all. \link{character}}

\item{start}{Either NA, meaning all available years from inputHBP data, or a character vector specifying start year in the form YYYY, e.g. 2019. The default and earliest allowable option is 2018. \link{character}}

\item{end}{Either NA, meaning all available years from inputHBP data after the start year, or a character vector specifying start year in the form YYYY, e.g. 2021. Data from the current calendar year are excluded since they would be incomplete. \link{character}}

\item{plotType}{Optional filter based on NEON plot type. Options are "all" (both tower and distributed plot types) or "tower". Defaults to "all". \link{character}}

\item{plotPriority}{NEON plots have a priority number in the event that not all plots are able to be sampled. The lower the number the higher the priority. The default is NA. \link{numeric}}
}
\value{
This function returns a list that includes biomass summary data frames.
'hbp_agb' summarizes above-ground herbaceous biomass for each record (units are g per m2).
'hbp_plot' summarizes above-ground live herbaceous biomass for each plot by year combination (units are both g per m2 and Megagrams per hectare).
'hbp_site' summarizes above-ground live herbaceous biomass for each site by year combination (units are both g per m2 and Megagrams per hectare).
}
\description{
Summarise aboveground herbaceous biomass. Biomass outputs can, if desired, be used in the follow-up estimateHerbProd productivity function.
Data inputs are "Herbaceous clip harvest" (DP1.10023.001) retrieved using the neonUtilities::loadByProduct() function (preferred), or data downloaded from the NEON Data Portal.
}
\details{
Input data can be filtered by site, date, plot type, and plot priority. Herbaceous biomass are scaled to an area basis at the hierarchical levels of clip cell, plot, and site.
}
\examples{
\dontrun{
# Obtain NEON Herbaceous clip harvest data
HbpDat <- neonUtilities::loadByProduct(dpID = "DP1.10023.001",
     site = "all",
     startdate = "2018-01",
     enddate = paste0(as.character(as.integer(format(Sys.Date(), "\%Y"))-1), "-12"),
     package = "basic", check.size = FALSE)
     
# If list is not in memory, load HBP list of dataframes from local file:
load('HbpDat.rds') # load NEON HBP portal data from a local file

# example with arguments at default values
scaleHerbMassOutputs <- scaleHerbMass(inputDataListHbp = HbpDat)

# example with inputDataListHbp = NA that triggers a fresh NEON portal data download
scaleHerbMassOutputs <- scaleHerbMass(inputDataListHbp = NA, site = "STEI")

# example with just a subset of inputDataListHbp data being utilized
scaleHerbMassOutputs <- scaleHerbMass(inputDataListHbp = HbpDat, start = 2022)

list2env(scaleHerbMassOutputs ,.GlobalEnv) # unlist all data frames
saveRDS(scaleHerbMassOutputs, 'scaleHerbMassOutputs.rds') # save all outputs locally
}
}
\author{
Samuel M Simkin \email{ssimkin@battelleecology.org} \cr
}
