% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPlantDiv.R
\name{stackPlantDiv}
\alias{stackPlantDiv}
\title{Stack NEON plant diversity data}
\usage{
stackPlantDiv(
  div_1m2Data = NA,
  div_10m2Data100m2Data = NA,
  totalSampledAreaFilter = NA
)
}
\arguments{
\item{div_1m2Data}{div_1m2Data table in NEON data produdct (Plant presence and percent cover, DP1.10058.001)}

\item{div_10m2Data100m2Data}{div_10m2Data100m2Data table in NEON data produdct (Plant presence and percent cover, DP1.10058.001)}

\item{totalSampledAreaFilter}{(integer, options are NA, 1, 10, 100, 400) The plot size for which data are returned. Default (NA) will return data for all plot sizes in the dataset. If you select a plot size, the function will filter the data returned to the desired plot size.}
}
\value{
This function returns a data frame
}
\description{
Use this function to aggregate NEON plant presence data to reflect plant species present at each plot scale.
}
\examples{
\dontrun{

# load neonUtilities
library (neonUtilities)

# get data
allDiv <- loadByProduct(
  dpID = "DP1.10058.001",
  site = "SRER",
  startdate = "2010-01",
  package = "basic",
  check.size = FALSE)

# stack all the div data
data_stacked <- stackPlantDiv(
  div_1m2Data = allDiv$div_1m2Data,
  div_10m2Data100m2Data = allDiv$div_10m2Data100m2Data)

# stack the data and filter to 10m plot
data_stacked_10m <- stackPlantDiv(
  div_1m2Data = allDiv$div_1m2Data,
  div_10m2Data100m2Data = allDiv$div_10m2Data100m2Data,
  totalSampledAreaFilter = 10)

}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Dave T Barnett \email{dbarnettl@battelleecology.org} \cr
}
