% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPlantPresence.R
\name{stackPlantPresence}
\alias{stackPlantPresence}
\title{Stack NEON Plant Presence and Percent Cover data}
\usage{
stackPlantPresence(
  inputDataList,
  input_1m2Data = NA,
  input_10m2Data100m2Data = NA,
  totalSampledAreaFilter = NA_integer_
)
}
\arguments{
\item{inputDataList}{A list object comprised of NEON Plant Presence and Percent Cover tables (DP1.10058.001) downloaded with the neonUtilities::loadByProduct() function. If list input is provided, the table input arguments must all be NA; similarly, if list input is missing, table inputs must be provided for the 'input_1m2Data' and 'input_10m2Data100m2Data' arguments.\link{list}}

\item{input_1m2Data}{The 'div_1m2Data' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be omitted. \link{data.frame}}

\item{input_10m2Data100m2Data}{The 'div_10m2Data100m2Data' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be omitted. \link{data.frame}}

\item{totalSampledAreaFilter}{The subplot or plot size for which data are returned (in meters squared). Default (NA) will return data for all subplot and plot sizes. If a valid filter integer is provided, the returned data are filtered to the desired subplot or plot size. Input options are NA, 1, 10, 100, 400. \link{integer}}
}
\value{
The following objects are returned as a list:
\itemize{
\item div_stackedPresence - A data frame that aggregates plant species lists across all subplots (or those subplots defined in 'totalSampledAreaFilter' argument), and for the entire plot where subplotID suffix is '400' (see Data Product documentation for subplot description).
\item variables - Units and definitions of novel variables created by the function that are not already defined in the Plant Presence and Percent Cover data product.
}
}
\description{
Aggregate occurrence data from the Plant Presence and Percent Cover data product to return all plant species present at the scale of each subplot sampled (1m2, 10m2, and 100m2) as well as the plot scale (400m2). Data downloaded from the NEON data portal report species in 10m2 and 100m2 subplots only if those species are not encountered in smaller-scale subplots, and a list of species present within the entire 400m2 scale plot is not provided. Data inputs are NEON Plant Presence and Percent Cover data (DP1.10058.001) retrieved with the neonUtilities::loadByProduct() function (preferred), data downloaded from the NEON Data Portal, or input data tables with an equivalent structure
}
\details{
Input data may be provided either as a list generated from the neonUtilities::loadByProduct() function or as individual tables. However, only list or table inputs are allowed (not a mix of both).
}
\examples{
\dontrun{

#   Obtain NEON Plant Presence and Percent Cover data; note that a token is required and may be obtained after creating a NEON user accounts get data
div <- neonUtilities::loadByProduct(
  dpID = "DP1.10058.001",
  site = "SRER",
  startdate = "2010-01",
  package = "basic",
  check.size = FALSE,
  token = "my_NEON_token"
  )

#   Stack the data by using the list returned by neonUtilities::loadByProduct as an input
data_stacked <- neonPlants::stackPlantPresence(
  inputDataList = div)

#   Send list of data using pipe
data_stacked <- div \%>\%
  neonPlants::stackPlantPresence()

#   Filter to 10m plots
data_stacked_10 <- div \%>\%
  neonPlants::stackPlantPresence(totalSampledAreaFilter = 10)

#   Filter to 10m plots outside of the stackPlantPresence function
data_stacked_10 <- div \%>\%
  neonPlants::stackPlantPresence() \%>\%
  dplyr::filter(totalSampledArea == 10)

#   Make your own list and stack the data
my_1m_data <- div$div_1m2Data
my_10_100m_data <- div$div_10m2Data100m2Data

data_stacked <- neonPlants::stackPlantPresence(
  inputDataList = list(
    div_1m2Data = my_1m_data,
    div_10m2Data100m2Data = my_10_100m_data))

#   Filter tables to a single plot, then stack the data
my_1m_data_SRER_43 <- my_1m_data \%>\%
  dplyr::filter(namedLocation == "SRER_043.basePlot.div")

my_10_100m_data_SRER_43 <- my_10_100m_data \%>\%
  dplyr::filter(namedLocation == "SRER_043.basePlot.div")

data_stacked_SRER_43 <- neonPlants::stackPlantPresence(
  inputDataList = list(
    div_1m2Data = my_1m_data_SRER_43,
    div_10m2Data100m2Data = my_10_100m_data_SRER_43))
}

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Dave T Barnett \email{dbarnettl@BattelleEcology.org} \cr
Eric Sokol \email{esokol@BattelleEcology.org} \cr
}
