% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeRootMass.R
\name{standardizeRootMass}
\alias{standardizeRootMass}
\title{Standardize NEON root dry mass data to current size and root status categories}
\usage{
standardizeRootMass(inputDataList, inputMass = NA)
}
\arguments{
\item{inputDataList}{A list object comprised of Plant Below Ground Biomass tables (DP1.10067.001) downloaded using the neonUtilities::loadByProduct() function (defaults to required). If list input is provided, the table input argument must be NA; similarly, if list input is missing, the 'inputMass' table input must be provided. \link{list}}

\item{inputMass}{The 'bbc_rootmass' table for the site x month combination(s) of interest (defaults to NA). If table input is provided, the 'inputDataList' argument must be omitted. \link{data.frame}}
}
\value{
A table containing root mass data for three sizeCategories (< 1mm, 1-2mm, and 2-10mm) and dryMass values pooled across previously utilized "live/dead" rootStatus categories. The output no longer contains the 'rootStatus' field, and any QA dryMass samples are averaged.
}
\description{
NEON root mass data collected prior to 2019 were sorted to four sizeCategories and were parsed into "live" and "dead" rootStatus categories. Beginning in 2019 the two smallest sizeCategories (< 0.5mm and 0.5-1mm) were combined into a single < 1mm sizeCategory and roots were no longer sorted to "live/dead". This function combines the older smallest two sizeCategories into the current < 1mm category and pools together roots previously categorized as "live" and "dead".

Data inputs are NEON Plant Belowground Biomass data (DP1.10067.001) in list format retrieved using the neonUtilities::loadByProduct() function (preferred), data tables downloaded from the NEON Data Portal, or input data tables with an equivalent structure.
}
\details{
Input data may be provided either as a list or as individual tables. However, if both list and table inputs are provided at the same time the function will error out.

NEON weighs a minimum of 5\% of samples a second time so that data users can estimate the uncertainty associated with different technicians weighing dried roots; QA samples of this nature are identified via qaDryMass == "Y". The function calculates the mean when QA masses exist and any 'remarks' are concatenated. Samples with Sampling Impractical values other than "OK" are removed prior to summarizing the input data.
}
\examples{
\dontrun{
#   Obtain NEON Plant Belowground Biomass data; note that a token is required and may be obtained after creating a NEON user account
bbcDF <- neonUtilities::loadByProduct(
dpID = "DP1.10067.001",
site = "all",
startdate = "2022-07",
enddate = "2022-08",
tabl = "all",
check.size = FALSE,
token = "my_NEON_token"
)

#   Standardize downloaded root data to current sizeCategories and rootStatus
df <- neonPlants::standardizeRootMass(
inputDataList = bbcDF,
inputMass = NA
)

}

}
\author{
Courtney Meier \email{cmeier@battelleecology.org} \cr
}
